### 2022-10-14 #4852 @editorialbot
[REVIEW]: lidarwind: A python package for retrieving wind profiles from Doppler lidar observations
**Submitting author:** @jdiasn (<a href="http://orcid.org/0000-0002-8488-8486">Dias Neto, José </a>)
**Repository:** https://github.com/jdiasn/lidarwind
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.2
