### 2022-10-14 #4852 @elbeejay
:wave: @tgoelles, @acocac, @bgailleton thank you all again for agreeing to review this submission. I believe most of the instructions on how to conduct a JOSS review are included in the comment at the top of this issue. The gist of the process is the following:

- Comment in this issue to tell the @editorialbot to make your checklist
- Review the software and associated paper (this includes installing, running, trying examples), and mark items off on your reviewer checklist. If anything does not work, or if items mentioned in the checklist are not complete, please open an issue in the **lidarSuit** repository itself, and describe the situation, when doing that, please include a link to **this** review issue so that we keep things connected
- Once you have completed your initial review of the submission, please comment here with your overall impression so that both @jdiasn and myself know what the next steps need to be

We are currently giving our reviewers 6 weeks to complete their reviews. To help ensure this review doesn't slip off of anyone's to-do list, I am going to ask the bot to send out reminders in this issue in 3 weeks time, and I will follow up at that point as well. If you need an extension or if something comes up that prevents you from doing this review please let me know. We try to maintain a transparent review process here at JOSS, and open communication certainly helps with that.

Thanks again all for being part of this process. JOSS relies on everyone's efforts to improve and publish scientific software. Please do not hesitate to reach out to me with any questions.
