### 2022-10-17 bab3ea3
Solving gdwon dependece (documentation) (#40)
