### 2022-10-19 3375220
Refactor: improving scan reference time assignment
