### 2022-10-19 ed838b3
Fix: read the docs dependence (#43)
