+++ lidarSuit/io.py
@@ -31,6 +31,7 @@ def open_sweep(file_name):
     ds = raw_data[f"/{sweep_group_name}"].to_dataset()
 
     assert "time_reference" in ds, "missing time_reference in input sweep"
+    # Guarantee that it is a valid datetime
     reference_time = pd.to_datetime(ds["time_reference"].values).isoformat()
     ds["time"].attrs["units"] = f"seconds since {reference_time}"
 
