### 2022-10-21 baec10c
Warning for getLidarData
