### 2022-10-25 0e8d40c
Deprecate get lidar data (#44)
* Adding minor inline info

* Warning for getLidarData

Co-authored-by: Dias Neto <jdiasn@gmail.com>
