### 2022-10-25 3bb00fc
refactor: tag gdown and lidarSuit versions
