### 2022-10-25 82a049a
Refactor/tag package version (#45)
* refactor: tag gdown and lidarSuit versions

* fixing sintaxe
