### 2022-10-25 jdiasn/lidarwind#46 @bgailleton
Minor comments [JOSS review]
Dear authors,

Before submitting my overall comments about the contribution, please find bellow minor comments regarding [the review for JOSS](https://github.com/openjournals/joss-reviews/issues/4852).

1. There are a couple of typos in the example notebooks (`docs/notebooks`). `quicklooks` and `turbulence_6beam_data` have path issues as well: files cannot be found or paths are hardcoded. As a more general note, I would suggest to add a bit more details in the notebooks (This is not my field of expertise and I find them a bit confusing).

2. The documentation's structure is a bit messy and could be clarified/improved. While it is easy to find clear guidelines to install the package and contribute, usage and general descriptions are harder to find. The content index points 3 times to the intro.

3. The documentation does not describes, even briefly, the API and only points to example notebooks for usage (which is already nice, but does not provide a lot of details).

4. The manuscript is well written and clear, I would only suggest to rephrase the very last sentence describing future additions. The JOSS paper (I believe) should remain general and up-to-date even in the case of future additions. Maybe rephrasing/extending this sentence around how it can be used to implement other standards, or to point to a dedicated page on the documentation describing the additions post-publication?

Boris
