### 2022-10-25 jdiasn/lidarSuit#46 @jdiasn
Dear @bgailleton , thanks for your comments. Those are valid points. We will address those and ping you in the respective PRs so you can keep track of it.
