### 2022-10-26 jdiasn/lidarSuit#47 @tgoelles
[JOSS Review]: test coverage too low
35% test coverage is too low. You don't need 100% but certainly more than 35%.
Please improve the coverage. I have seen that you have it in the makefile.

https://github.com/openjournals/joss-reviews/issues/4852
