### 2022-10-26 jdiasn/lidarSuit#48 @tgoelles
[JOSS review]: follow PEP 8
There is a lot of CamelCase used where it shouldn't.

If you run pylint * in the lidarSuit directory you get a score.
Currently it's 4.07

This is mostly caused by: doesn't conform to snake_case naming style

I suggest increasing the test coverage first and then tackle the names. I guess there is tool out there which does most of the work for you.

https://github.com/openjournals/joss-reviews/issues/4852
