+++ environment.yml
@@ -13,7 +13,7 @@ dependencies:
  - netCDF4=1.5.8
  - matplotlib=3.5.1
  - click=8.1.2
- - xarray-datatree=0.0.6
+ - xarray-datatree=0.0.9
  - pip
  - pip:
      - gdown
