+++ pyproject.toml
@@ -40,7 +40,7 @@ dependencies = [
   "matplotlib>=3.4.3",
   "click>=8.1.2",
   "gdown~=4.5.1",
-  "xarray-datatree~=0.0.6"
+  "xarray-datatree==0.0.9"
 ]
 
 [project.optional-dependencies]
