### 2022-10-26 jdiasn/lidarwind#49 @tgoelles
[JOSS Review]: package name not following PEP8 and might be misleading
Similar to issue #48

Consider renaming the package. You shouldn't use camel case there.
Also I think that lidarSuit is not really a suitable name, since lidar is so much more and wind lidar is just a smaller part if it.
The name should be short and meaningful.

I suggest :
windcube or windlidar

In some text you write about lidar_suit. Please, keep it consistent.

This is just a suggestion and I think you would benefit from it since its draws in the right people.

https://github.com/openjournals/joss-reviews/issues/4852
