### 2022-10-26 d214dd4
Adding pytest-cov into test optionals
