+++ pyproject.toml
@@ -45,7 +45,7 @@ dependencies = [
 [project.optional-dependencies]
 plots = ["matplotlib>=3.4.3"]
 dev = ["sphinx_rtd_theme==0.4.3"]
-test = ["pytest", "gdown~=4.5.1"]
+test = ["pytest", "gdown~=4.5.1", "pytest-cov>=3.0"]
 
 [project.urls]
 homepage = "https://github.com/jdiasn/lidarSuit"
