### 2022-10-26 jdiasn/lidarSuit#47 @castelao
Indeed. Thanks for pointing that out. We will work on it.
