### 2022-10-26 4a2245b
Defining cache key
