+++ .github/workflows/ci.yml
@@ -50,6 +50,7 @@ jobs:
       with:
         python-version: "3.10"
         cache: "pip"
+        cache-dependency-path: 'pyproject.toml'
 
     - name: Coverage report
       run: |
