### 2022-10-26 jdiasn/lidarSuit#47 @jdiasn
Dear @tgoelles, thank you for your comment! We agree with you that the test coverage is low. We will be working to improve it.
