### 2022-10-26 jdiasn/lidarSuit#49 @jdiasn
@tgoelles, thank you for pointing this out. We are going to remove the camel case and re-think the package name.
