### 2022-10-26 jdiasn/lidarSuit#48 @jdiasn
@tgoelles, I agree with you on it. I used CamelCase a lot. We will improve it and use the snake_case instead.
