### 2022-10-27 35a2529
Code coverage (#50)
* Reporting coverage

* Adding pytest-cov into test optionals

* Adding temporary badge to be updated later

* Defining cache key

* Updating badge
