+++ .github/workflows/ci.yml
@@ -37,3 +37,29 @@ jobs:
     - name: Test with pytest
       run: |
         pytest tests
+
+  coverage:
+    runs-on: ubuntu-latest
+
+    steps:
+    - name: checkout lidarSuit
+      uses: actions/checkout@v2
+
+    - name: Set up Python
+      uses: actions/setup-python@v4
+      with:
+        python-version: "3.10"
+        cache: "pip"
+        cache-dependency-path: 'pyproject.toml'
+
+    - name: Coverage report
+      run: |
+        python -m pip install --upgrade pip
+        pip install -e .[test]
+        pytest --cov=./ --cov-report=xml
+
+    - name: Upload coverage to Codecov
+      uses: codecov/codecov-action@v3
+      with:
+        fail_ci_if_error: true
+        verbose: true
