### 2022-10-27 7cbab24
Adding io.py (#39)
* Adding io.py

* BUG fix to run xarray 2022.06.0 (#38)

* Adding test IO

* Fix: fixing compatibility xarray issue

* Adding vertical data check

* updating sample data path

* Fix: correcting gdwon dependence

* Removing old requirements

* Fix: correcting tag dependence

* Updating xarray dependence

* Feat: functionalities to download data samples

* Fixing dependence

* Solving gdwon dependece (documentation) (#40)

* Refactor/big file compatibility (#41)

* Solving gdwon dependece (documentation)

* Fix: gdown dependence

* Refactor/big file compatibility (#42)

* Solv
[…]
