+++ environment.yml
@@ -6,7 +6,7 @@ channels:
 
 dependencies:
  - python=3.8
- - xarray=2022.3.0
+ - xarray=2022.6.0
  - xrft=0.4.1
  - pandas=1.4.2
  - numpy=1.22.3
