### 2022-10-28 8fb71c3
setting Read the Docs conda channels (#52)
