+++ docs/environment.yml
@@ -1,5 +1,9 @@
 name: lidarSuit
 
+channels:
+  - defaults
+  - conda-forge
+
 dependencies:
  - python=3.8
  - numpy
