### 2022-10-30 jdiasn/lidarwind#53 @acocac
[JOSS review] Comments
* LIDAR Suit version: 0.1.6
* Python version: 3.8
* Operating System: Mac OS

#
Dear authors,

Thanks for the submission of the paper. I've found very interesting the contribution, proposed features and efforts to building upon existing open-source python packages for handling N-dimensional datasets e.g. `xarray`. Please find below general and minor comments. Also, I've posted my overall comments in the paper [REVIEW](https://github.com/openjournals/joss-reviews/issues/4852).

# General
- ~~The documentation would benefit from adding more details of `lidarSuit`. There is only a short description of how to generate a config file, however notebooks describe key functions i.e. `sixBeamMethod`, `dbsOperations`, etc.~~
- ~~The `notebook` folder is located within the `docs` dir. I would suggest moving it to the root dir. Also, you can create a section of rendered notebooks in the documentation.~~
-  ~~I think the paper is clear and well written. It would be great to add a diagram of the architecture of the library (flowchart and class hierarchy of the package). You could point into the same diagram future developments e.g. computing the Reynold stress for the VAD and DBS. Also, for non-experts it would be helpful to include some visualisation outputs with a brief interpretation.~~
- Binder link doesn’t work (Failed to connect to event stream).
- ~~Typos in README, luidarSuit, 6-bean.~~
- Add a short and more accessible How to Contribute in README.

# Others (optional):
- ~~The package name should be changed as it focuses only on wind lidar.~~
- ~~Google drive is not the most reliable storage for sample data. I suggest using Zenodo or another platform alike. There are some handy python open-source packages to handle files in Zenodo. For instance, the [pooch](https://github.com/fatiando/pooch) package enforces that everyone running the code has the same version of the data files (through the SHA256 hashes).~~
