### 2022-10-30 #4852 @acocac
In this paper authors present a new Python package, `lidarSuit`, whose main goal is to retrieve wind profiles from Doppler lidar systems. The paper is well-written with a good introduction and a clear statement of need. Authors claim the key difference to other commonly-used software is the support of 6-beam observations using the Fast Fourier Wind Vector Algorithm plus other functionalities i.e. computing the Reynold stress tensor. While the source code contains some useful notebooks examples, the documentation needs to be improved. In addition, I miss further details of other software (names at least) and their corresponding citations.

I strongly believe that `lidarSuit` has a great potential and could become a very useful Python package once the authors better describe the package and how to use it, both in the repository and in the paper. I have opened an issue in the [source repository](https://github.com/jdiasn/lidarSuit/issues/53) with some general and minor comments. Feel free to decide how to proceed.
