### 2022-10-31 #4852 @elbeejay
Thanks @tgoelles, @bgailleton, and @acocac for taking the time to review this submission and open up relevant review issues in the `lidarSuit` repository. From what I can tell, it does not appear that any of the review comments are going to be barriers to future publication; most appear to be related to improving the clarity of the code and documentation. @jdiasn, it appears that you have seen some of these comments and have provided responses and are actively working on making changes so that is great. Please let us know once you believe the revisions you've made satisfy the comments, and be sure to tag the reviewers and myself.

Thanks all!
