### 2022-10-31 jdiasn/lidarSuit#53 @jdiasn
@acocac Thank you for your review! Your suggestions were very interesting and important for the further development of the package. We will be working to incorporate those suggestions and will let you know as soon as they are implemented.
