### 2022-11-02 #4852 @bgailleton
@elbeejay I confirm that from my side, I consider [my issue](https://github.com/jdiasn/lidarSuit/issues/46) pretty minor and I recommend publication as soon as the documentation and examples improve. The contribution presents a `python` package gathering and combining multiple methods to retrieve wind profiles from LiDAR datasets. It bringa to the community methods that were not as easily accessible before and is susceptible to improve scientific research in the field. I cannot really judge the academic value as this is a bit out of my field of expertise, but the code is well packaged, thoroughly commented, easy to install and delivers what the manuscript promise.
