### 2022-11-07 jdiasn/lidarwind#54 @tgoelles
[JOSS Review]: minor issues
# Minor issues

* A lot of commented code needs to be deleted. That's what version control is for.
* look up f-strings. Makes the code more readable. You don't need to change your code, but just a hint for the future

## Documenation
* you have a docstring in the methods. Let Sphinx generate the html documentation from it and include it in readthedocs
* add the API reference to the sidebar menu (Contents)

## jupyter notebooks

* why use uncomment in cells in the jupyter notebooks? Just delete the output cells before saving. You could use nbstripout for that
* quicklooks.ipynb did not work: FileNotFoundError
* turbulence_6beam_data.ipyn did not work: FileNotFoundError
* you should add the notebooks to the test suite:  pytest --cov-append --current-env --nbval-lax xxxxx.ipynb (this only checks if the cells evaluated without error)

## Paper

* use Python not python in the paper
* cite NetCDF: <https://www.unidata.ucar.edu/support/help/MailArchives/netcdf/msg00087.html>

https://github.com/openjournals/joss-reviews/issues/4852
