### 2022-11-09 jdiasn/lidarSuit#54 @jdiasn
@tgoelles Thank you again for your suggestions! They are very interesting and cover points that I had missed during the development.
