+++ .pre-commit-config.yaml
@@ -4,11 +4,16 @@ default_language_version:
     python: python3.9
 repos:
 -   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v3.2.0
+    rev: v4.3.0
     hooks:
+    -   id: check-ast
+    -   id: check-docstring-first
+    -   id: check-merge-conflict
+    -   id: check-yaml
+    -   id: check-toml
+    -   id: debug-statements
     -   id: trailing-whitespace
     -   id: end-of-file-fixer
-    -   id: check-yaml
 -   repo: https://github.com/psf/black
     rev: 22.3.0
     hooks:
