### 2022-11-14 b9fdae3
Adding pyupgrade to pre-commit
