+++ .pre-commit-config.yaml
@@ -14,6 +14,10 @@ repos:
     -   id: debug-statements
     -   id: trailing-whitespace
     -   id: end-of-file-fixer
+-   repo: https://github.com/asottile/pyupgrade
+    rev: v2.38.2
+    hooks:
+    -   id: pyupgrade
 -   repo: https://github.com/psf/black
     rev: 22.3.0
     hooks:
