+++ .pre-commit-config.yaml
@@ -14,6 +14,7 @@ repos:
     -   id: debug-statements
     -   id: trailing-whitespace
     -   id: end-of-file-fixer
+    -   id: check-added-large-files
 -   repo: https://github.com/asottile/pyupgrade
     rev: v2.38.2
     hooks:
