+++ .pre-commit-config.yaml
@@ -19,6 +19,7 @@ repos:
     rev: v2.38.2
     hooks:
     -   id: pyupgrade
+        args: [--py38-plus]
 -   repo: https://github.com/psf/black
     rev: 22.8.0
     hooks:
