### 2022-11-14 23a6d3f
Configuring flake8
