+++ setup.cfg
@@ -2,6 +2,15 @@
 universal = 1
 
 [flake8]
-exclude = docs
+max-complexity = 22
+max-line-length = 79
+ignore = E203, W503, B011
+exclude =
+    .git,
+    __pycache__,
+    build,
+    dist,
+    docs/conf.py
+
 [tool:pytest]
 collect_ignore = ['setup.py']
