### 2022-11-15 e00cd90
Moving pytest configuration to pyproject.toml
