+++ setup.cfg
@@ -11,6 +11,3 @@ exclude =
     build,
     dist,
     docs/conf.py
-
-[tool:pytest]
-collect_ignore = ['setup.py']
