+++ pyproject.toml
@@ -67,3 +67,6 @@ line-length = 79
 [tool.setuptools_scm]
 write_to = "lidarSuit/version.py"
 git_describe_command = "git describe --dirty --tags --long --match 'v*' --first-parent"
+
+[tool.pytest.ini_options]
+testpaths = ["tests"]
