### 2022-11-15 f715c19
Improving code quality checkers (#56)
* Updating pre-commit-hooks

* Adding pyupgrade to pre-commit

* Updating black @ pre-commit

* Check for large files

* pyupgrade with --py38-plus

* Adding flake8 to pre-commit

* Extending dev requirements

* Configuring flake8

* Moving pytest configuration to pyproject.toml
