### 2022-11-15 cbb5516
Adding a notebook test
