### 2022-11-15 1df2420
Sintaxe check
