+++ .github/workflows/ci.yml
@@ -57,7 +57,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install -e .[test]
         pytest --cov=./ --cov-report=xml
-        pytest --cov=./docs/notebooks/ --cov-append --cov-report=xml turbulence_6beam_data.ipynb
+        pytest --cov=./ --cov-append --cov-report=xml --nbval-lax docs/notebooks/turbulence_6beam_data.ipynb
 
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v3
