### 2022-11-15 8f7425a
Adding inital notebook tests (#57)
* Refactor: removing notebooks cell's output

* Adding a notebook test

* Adding new test dependecy

* Sintaxe check

* Adding test dependecy: pooch

* Adding a notebook test
