+++ .github/workflows/ci.yml
@@ -57,6 +57,8 @@ jobs:
         python -m pip install --upgrade pip
         pip install -e .[test]
         pytest --cov=./ --cov-report=xml
+        pytest --cov=./ --cov-append --cov-report=xml --current-env --nbval-lax docs/notebooks/turbulence_6beam_data.ipynb
+        pytest --cov=./ --cov-append --cov-report=xml --current-env --nbval-lax docs/notebooks/merging_6beam_data.ipynb
 
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v3
