+++ lidarSuit/visualization.py
@@ -1,5 +1,4 @@
 import os
-import xarray as xr
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
