+++ lidarSuit/visualization.py
@@ -4,8 +4,6 @@ import pandas as pd
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
 
-# import matplotlib as mpl
-
 
 class plotSettings:
     def __init__(self, mpl, style="dark_background"):
