### 2022-11-17 8090aed
Removing commented filtering
@jdiasn, could you double check this please? This is exactly the
filters.filtering, so nothing lost removing these commented block,
right?
