+++ lidarSuit/visualization.py
@@ -4,6 +4,7 @@ import pandas as pd
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
 
+from .filters import filtering
 
 class plotSettings:
     def __init__(self, mpl, style="dark_background"):
