### 2022-11-17 65f7043
A simple function to filter based on the status
