### 2022-11-17 cf5e44f
Prottype of filter_snr()
