### 2022-11-18 78c4bcf
Cleaning and fixing visualization module (#58)
* Removing unused import

* Removing commented code

* Removing commented filtering

@jdiasn, could you double check this please? This is exactly the
filters.filtering, so nothing lost removing these commented block,
right?

* Upgrading print syntax (with pyupgrade)

* Missing import filtering

It looks like you previously moved it to another module and forgot to
import it after.

* PEP8
