+++ lidarSuit/visualization.py
@@ -1,11 +1,10 @@
 import os
-import xarray as xr
 import numpy as np
 import pandas as pd
 import matplotlib.pyplot as plt
 import matplotlib.dates as mdates
 
-# import matplotlib as mpl
+from .filters import filtering
 
 
 class plotSettings:
@@ -55,48 +54,6 @@ class plotSettings:
         return plot
 
 
-# class filtering:
-
-#     def __init__(self, data):
-
-#         self.data = data
-
-#         return None
-
-#     def getVerticalObsComp(self, variable, snr=False, status=True):
-
-#         tmpData = self.data[variable]
-
-#         if status:
-#             tmpData = tmpData.where(self.data.radial_wind_speed_status90==1)
-
-
-#         if snr != False:
-#             tmpData = tmpData.where(self.data.cnr90 > snr)
-
-
-#         tmpData = tmpData.where(self.data.elevation==90, drop=True)
-
-
-#         return tmpData
-
-#     def getRadialObsComp(self, variable, azm, snr=False, status=True):
-
-#         tmpData = self.data[variable]
-
-#         if status:
-#             tmpData = tmpData.where(self.data.radial_wind_speed_status==1)
-
-
-#         if snr != False:
-#             tmpData = tmpData.where(self.data.cnr > snr)
-
-
-#         tmpData = tmpData.where((self.data.elevation!=90) & (self.data.azimuth==azm), drop=True)
-
-#         return tmpData
-
-
 class visualizer:
     def __init__(self, data):
 
@@ -135,7 +92,7 @@ class visualizer:
             tmpData = filtering(self.data).getVerticalObsComp(varName)
 
             if namePrefix:
-                strName = "{0}_{1}".format(
+                strName = "{}_{}".format(
                     namePrefix, tmpData.attrs["standard_name"]
                 )
 
@@ -222,13 +179,13 @@ class visualizer:
         plt.grid(b=True)
         plt.ylim(0, 12e3)
         plt.xlim(pd.to_datetime(selDay.strftime("%Y%m%d")), maxTime)
-        plt.title("elv: {0}, azm: {1}".format(elv, azm))
+        plt.title(f"elv: {elv}, azm: {azm}")
 
         if plotID == "hor_wind_dir":
             plot.colorbar.set_ticks(np.linspace(0, 360, 9))
 
         if save:
-            fileName = "{0}_{1}.png".format(selDay.strftime("%Y%m%d"), plotID)
+            fileName = "{}_{}.png".format(selDay.strftime("%Y%m%d"), plotID)
             outputFileName = os.path.join(figPath, fileName)
             print(outputFileName)
             plt.savefig(outputFileName, bbox_inches="tight")
@@ -275,10 +232,10 @@ class visualizer:
             axes[axN].grid(b=True)
             axes[axN].set_ylim(0, 12e3)
             axes[axN].set_xlim(pd.to_datetime(tmpData.time[0].values), maxTime)
-            axes[axN].set_title("elv: {0}, azm: {1}".format(elv, i))
+            axes[axN].set_title(f"elv: {elv}, azm: {i}")
 
         if save:
-            fileName = "{0}_{1}.png".format(selDay.strftime("%Y%m%d"), plotID)
+            fileName = "{}_{}.png".format(selDay.strftime("%Y%m%d"), plotID)
             outputFileName = os.path.join(figPath, fileName)
             print(outputFileName)
             plt.savefig(outputFileName, bbox_inches="tight")
