### 2022-11-21 3fe042c
Prototype of pre-processing
@jdiasn, could you point me to an example of elevation=90 to guide this
development, please? It is not clear to me if we can have a 90 degrees
case in the same file with a non 90 degres elevation.
