### 2022-11-27 b9c0dd6
refactor: adding the dbs_notebook to test routine
