+++ .github/workflows/ci.yml
@@ -60,6 +60,7 @@ jobs:
         pytest --cov=./ --cov-append --cov-report=xml --current-env --nbval-lax docs/notebooks/turbulence_6beam_data.ipynb
         pytest --cov=./ --cov-append --cov-report=xml --current-env --nbval-lax docs/notebooks/merging_6beam_data.ipynb
         pytest --cov=./ --cov-append --cov-report=xml --current-env --nbval-lax docs/notebooks/dbs_scans.ipynb
+        pytest --cov=./ --cov-append --cov-report=xml --current-env --nbval-lax docs/notebooks/reading_long_dbs.ipynb
 
 
     - name: Upload coverage to Codecov
