### 2022-11-28 fb95b84
refactor: updating the dbs_scans notebook (#61)
* refactor: Adding a introduction

* refactor: caching sample data using pooch

* refactor: adding the dbs_notebook to test routine

* refactor: upgrading 6-beam md5

* refactor: adding an introduction to the long_dbs

* Upgrading long_dbs to cache data using pooch

* refactor: adding reading_long_dbs to to tests

* improving the description
