+++ .github/workflows/ci.yml
@@ -59,6 +59,9 @@ jobs:
         pytest --cov=./ --cov-report=xml
         pytest --cov=./ --cov-append --cov-report=xml --current-env --nbval-lax docs/notebooks/turbulence_6beam_data.ipynb
         pytest --cov=./ --cov-append --cov-report=xml --current-env --nbval-lax docs/notebooks/merging_6beam_data.ipynb
+        pytest --cov=./ --cov-append --cov-report=xml --current-env --nbval-lax docs/notebooks/dbs_scans.ipynb
+        pytest --cov=./ --cov-append --cov-report=xml --current-env --nbval-lax docs/notebooks/reading_long_dbs.ipynb
+
 
     - name: Upload coverage to Codecov
       uses: codecov/codecov-action@v3
