### 2022-12-02 2274065
re-formating strings: dataAttributesL1
