### 2022-12-02 de18c81
dataOperator: re-formating string, flake8
