### 2022-12-02 088c653
Temporary removing notebooks' tests
