### 2022-12-02 932349d
refactor: updating ci (#63)
* Updating CI to use setup-python v4

* Temporary removing notebooks' tests
