### 2022-12-02 ab70ecf
Pre-processing sweep
@jdiasn, based on what we talked about, does it make sense?
