### 2022-12-05 137ed97
Updating the initial plotting time
