### 2022-12-06 364596b
Rotines to set max and min time
