### 2022-12-06 466e05b
Upgrading quicklooks example (#64)
* Updating the initial plotting time

* Adding parameters to adjust time limites

* Rotines to set max and min time

* Updating plot limits

* Adding an introduction

* Splitting the quicklooks example into small steps
