### 2022-12-06 6794693
Adjusting style
Using pre-commit to adjust style. Changes limited to style.
