+++ joss/paper.bib
@@ -45,7 +45,7 @@
 
 @dataset{diasneto2022a,
   author       = {Dias Neto, José},
-  title        = {{The Tracing Convective Momentum Transport in 
+  title        = {{The Tracing Convective Momentum Transport in
                    Complex Cloudy Atmospheres Experiment - Level 1}},
   month        = {jul},
   year         = {2022},
