+++ lidarSuit/__init__.py
@@ -1,9 +1,8 @@
-
 """Top-level package for lidarSuit package"""
 
 __author__ = "José Dias Neto"
 __email__ = "jdiasn@gmail.com"
-__affiliation__="Delft University of Technology"
+__affiliation__ = "Delft University of Technology"
 
 
 from pkg_resources import get_distribution, DistributionNotFound
