### 2022-12-08 c17aeb6
Adding NetCDF reference/citation
