### 2022-12-08 40bfc6d
Upgrade/paper small corrections (#67)
* Correcting typo

* Adding NetCDF reference/citation
