+++ joss/paper.bib
@@ -45,7 +45,7 @@
 
 @dataset{diasneto2022a,
   author       = {Dias Neto, José},
-  title        = {{The Tracing Convective Momentum Transport in 
+  title        = {{The Tracing Convective Momentum Transport in
                    Complex Cloudy Atmospheres Experiment - Level 1}},
   month        = {jul},
   year         = {2022},
@@ -72,3 +72,13 @@
    year = {2017},
    url={http://resolver.tudelft.nl/uuid:a659654b-e76a-4513-a656-ecad761bdbc8}
     }
+
+@article{rew1990,
+  author={Rew, R. and Davis, G.},
+  journal={IEEE Computer Graphics and Applications},
+  title={NetCDF: an interface for scientific data access},
+  year={1990},
+  volume={10},
+  number={4},
+  pages={76-82},
+  doi={10.1109/38.56302}}
