### 2022-12-09 72d96f7
Adjusting style (#65)
* Adjusting style

Using pre-commit to adjust style. Changes limited to style.

* missed formated string

* typo

Co-authored-by: Dias Neto <jdiasn@gmail.com>
