+++ lidarSuit/windPropRetrieval.py
@@ -362,9 +362,7 @@ class getWindProperties5Beam:
         """
 
         self.logger.info(
-            "replacing the gate_index coordinate by range: {0}".format(
-                comp.name
-            )
+            f"replacing the gate_index coordinate by range: {comp.name}"
         )
 
         comp = comp.rename({"gate_index": "range"})
