+++ tests/test_lidar_suit.py
@@ -7,6 +7,7 @@ import pytest
 from click.testing import CliRunner
 
 import lidarSuit
+
 # from lidarSuit import cli
 
 
@@ -27,13 +28,13 @@ def test_content(response):
 
 
 # @pytest.skip('we do not have a command line yet')
-@pytest.mark.skip('we do not have a command line yet')
+@pytest.mark.skip("we do not have a command line yet")
 def test_command_line_interface():
     """Test the CLI."""
     runner = CliRunner()
     result = runner.invoke(cli.main)
     assert result.exit_code == 0
-    assert 'lidarSuit.cli.main' in result.output
-    help_result = runner.invoke(cli.main, ['--help'])
+    assert "lidarSuit.cli.main" in result.output
+    help_result = runner.invoke(cli.main, ["--help"])
     assert help_result.exit_code == 0
-    assert '--help  Show this message and exit.' in help_result.output
+    assert "--help  Show this message and exit." in help_result.output
