### 2022-12-09 e80e099
Naming new function as suggested by @jdiasn
