### 2022-12-09 4be35a4
replacing tmpHeight by tmp_height
