### 2022-12-09 a608b60
replacing removeCloud by remove_cloud
