### 2022-12-09 459dde2
replacing tmpCeiloHgt by tmp_ceilo_hgt
