### 2022-12-09 d3f45a3
replacing lowestBeta by lowest_beta
