### 2022-12-09 0f9cf0e
Moving support functions to its own module
