### 2022-12-09 96e042a
replacing positiveBeta by positive_beta
