### 2022-12-09 a8e8aa0
replacing getInterpInterfHeight by get_interp_interf_height
