### 2022-12-09 4d4f4b4
replacing getHeightInterface by get_height_iterface
