### 2022-12-09 95264c8
replacing getNoiseFreeBeta by get_noise_free_beta
