### 2022-12-09 0d333ba
replacing lidar != None by lidar is not None
