### 2022-12-09 a028872
replacing interpInterfHeight by interp_interf_height
