### 2022-12-09 8e42768
replacing interfHeight by interf_height
