### 2022-12-09 562f90f
replacing interpInterfHeight by interp_interf_height
