### 2022-12-09 4215ef6
replacing interfHeight by interf_height
