### 2022-12-09 87d049d
replacing noiseFreeBeta by noise_free_beta
