### 2022-12-09 7d83f24
replacing windCubeCloudRemoval by wind_cube_cloud_removal
