### 2022-12-09 50cd66c
replacing wind_cube_cloud_removal by WindCubeCloudRemoval
