### 2022-12-09 5407d9d
New support function lidarsuitrc()
