### 2022-12-09 aa4c678
New support function for tests: sample_data()
