### 2022-12-09 4fec07c
replacing anomStd by anom_std
