+++ lidarSuit/filters.py
@@ -241,7 +241,7 @@ class secondTripEchoFilter:
             & (self.lidar.dataTransf.time < self.endTime)
         )
 
-        anomStd = tmpAnom.std(dim=["time", "range", "elv"])
+        anom_std = tmpAnom.std(dim=["time", "range", "elv"])
 
         # Cross check if this commented part is still needed for
         # for filter the slanted profiles
@@ -249,14 +249,14 @@ class secondTripEchoFilter:
         # tmpNoCloud = self.lidar.dataTransf.where(self.timeCloudMask == 0).copy()
         # tmpCloud = self.lidar.dataTransf.where(self.timeCloudMask == 1).copy()
 
-        # tmpCloud = tmpCloud.where(np.abs(self.dataAnom) < self.nStd * anomStd)
+        # tmpCloud = tmpCloud.where(np.abs(self.dataAnom) < self.nStd * anom_std)
 
         # tmp_clean_data = tmpNoCloud.copy()
         # tmp_clean_data.values[np.isfinite(tmpCloud)] = tmpCloud.values[np.isfinite(tmpCloud)]
 
         tmp_clean_data = self.lidar.dataTransf.copy()
         tmp_clean_data = tmp_clean_data.where(
-            np.abs(self.dataAnom) < self.nStd * anomStd
+            np.abs(self.dataAnom) < self.nStd * anom_std
         )
         self.lidar.dataTransf.values = tmp_clean_data.values
 
@@ -271,19 +271,19 @@ class secondTripEchoFilter:
             & (self.lidar.dataTransf90.time < self.endTime)
         )
 
-        anomStd = tmpAnom.std(dim=["time", "range90"])
+        anom_std = tmpAnom.std(dim=["time", "range90"])
 
         # tmpNoCloud = self.lidar.dataTransf90.where(self.timeCloudMask == 0).copy()
         # tmpCloud = self.lidar.dataTransf90.where(self.timeCloudMask == 1).copy()
 
-        # tmpCloud = tmpCloud.where(np.abs(self.dataAnom90) < self.nStd * anomStd)
+        # tmpCloud = tmpCloud.where(np.abs(self.dataAnom90) < self.nStd * anom_std)
 
         # tmp_clean_data = tmpNoCloud.copy()
         # tmp_clean_data.values[np.isfinite(tmpCloud)] = tmpCloud.values[np.isfinite(tmpCloud)]
 
         tmp_clean_data = self.lidar.dataTransf90.copy()
         tmp_clean_data = tmp_clean_data.where(
-            np.abs(self.dataAnom90) < self.nStd * anomStd
+            np.abs(self.dataAnom90) < self.nStd * anom_std
         )
 
         self.lidar.dataTransf90.values = tmp_clean_data.values
