### 2022-12-09 2014020
replacing tmpSelData by tmp_sel_data
