### 2022-12-09 74c88db
replacing calMeanAndAnomSlant by cal_mean_and_anom_slant
