### 2022-12-09 dbe002f
fix: Dirty fix to download sample data before refactoring
