### 2022-12-09 b3bb765
replacing strH by str_h
