### 2022-12-09 0c32670
replacing getTimeEdges by get_time_edges
