### 2022-12-09 926babd
replacing getVerticalObsComp by get_vertical_obs_comp
