### 2022-12-09 3a65c5b
replacing snr != False by snr is not False
