### 2022-12-09 6ffe75e
replacing getRadialObsComp by get_radial_obs_comp
