### 2022-12-09 55452f5
replacing secondTripEchoFilter by SecondTripEchoFilter
