### 2022-12-09 6a074a5
replacing nStd by n_std
