+++ lidarSuit/filters.py
@@ -198,7 +198,7 @@ class SecondTripEchoFilter:
 
         sel_time = pd.to_datetime(self.lidar.dataTransf.time.values[0])
         sel_time = sel_time.strftime("%Y%m%d")
-        self.startTime = pd.to_datetime(f"{sel_time} {str_h}")
+        self.start_time = pd.to_datetime(f"{sel_time} {str_h}")
         self.endTime = pd.to_datetime(f"{sel_time} {end_h}")
 
 
@@ -237,7 +237,7 @@ class SecondTripEchoFilter:
         """
 
         tmp_anom = self.dataAnom.where(
-            (self.lidar.dataTransf.time > self.startTime)
+            (self.lidar.dataTransf.time > self.start_time)
             & (self.lidar.dataTransf.time < self.endTime)
         )
 
@@ -267,7 +267,7 @@ class SecondTripEchoFilter:
         """
 
         tmp_anom = self.data_anom_90.where(
-            (self.lidar.dataTransf90.time > self.startTime)
+            (self.lidar.dataTransf90.time > self.start_time)
             & (self.lidar.dataTransf90.time < self.endTime)
         )
 
