+++ lidarSuit/utilities.py
@@ -24,9 +24,9 @@ class util:
         start_time = pd.to_datetime(f"{start} 00:00:00")
 
         end = (selDay + pd.to_timedelta(1, "D")).strftime("%Y%m%d")
-        endTime = pd.to_datetime(f"{end} 00:00:00")
+        end_time = pd.to_datetime(f"{end} 00:00:00")
 
-        timeBins = pd.date_range(start_time, endTime, freq=freq)
+        timeBins = pd.date_range(start_time, end_time, freq=freq)
 
         return timeBins
 
