### 2022-12-09 ba0ce53
replacing dataMean by data_mean
