### 2022-12-10 dd2ac6d
Adjusting camel case on filtering module (#68)
* replacing tmpHeight by tmp_height

* replacing tmpValues by tmp_values

* replacing removeCloud by remove_cloud

* replacing tmpCeiloHgt by tmp_ceilo_hgt

* replacing lowestBeta by lowest_beta

* replacing positiveBeta by positive_beta

* replacing getInterpInterfHeight by get_interp_interf_height

* formated string

* replacing getHeightInterface by get_height_iterface

* replacing getNoiseFreeBeta by get_noise_free_beta

* replacing lidar != None by lidar is not None

* replacing interpInterfHeight90 by interp_interf_height_90

* replacing int
[…]
