+++ lidarSuit/filters.py
@@ -6,7 +6,7 @@ import pandas as pd
 import numpy as np
 
 
-class filtering:
+class Filtering:
     """SNR and Status filter
 
     It uses the carrier to noise ratio (SNR) and status
@@ -24,7 +24,7 @@ class filtering:
 
         self.data = data
 
-    def getVerticalObsComp(self, variable, snr=False, status=True):
+    def get_vertical_obs_comp(self, variable, snr=False, status=True):
         """Vertical data filter
 
         It uses the SNR and status variables to filter out
@@ -46,25 +46,25 @@ class filtering:
 
         Returns
         -------
-        tmpData : xarray.DataSet
+        tmp_data : xarray.DataSet
             an instance of the variable filtered using
             SNR or status variable
 
         """
 
-        tmpData = self.data[variable]
+        tmp_data = self.data[variable]
 
         if status:
-            tmpData = tmpData.where(self.data.radial_wind_speed_status90 == 1)
+            tmp_data = tmp_data.where(self.data.radial_wind_speed_status90 == 1)
 
-        if snr != False:
-            tmpData = tmpData.where(self.data.cnr90 > snr)
+        if snr is not False:
+            tmp_data = tmp_data.where(self.data.cnr90 > snr)
 
-        tmpData = tmpData.where(self.data.elevation == 90, drop=True)
+        tmp_data = tmp_data.where(self.data.elevation == 90, drop=True)
 
-        return tmpData
+        return tmp_data
 
-    def getRadialObsComp(self, variable, azm, snr=False, status=True):
+    def get_radial_obs_comp(self, variable, azm, snr=False, status=True):
         """Slanted data filter
 
         It uses the SNR and status variables to filter out
@@ -93,23 +93,23 @@ class filtering:
 
         """
 
-        tmpData = self.data[variable]
+        tmp_data = self.data[variable]
 
         if status:
-            tmpData = tmpData.where(self.data.radial_wind_speed_status == 1)
+            tmp_data = tmp_data.where(self.data.radial_wind_speed_status == 1)
 
-        if snr != False:
-            tmpData = tmpData.where(self.data.cnr > snr)
+        if snr is not False:
+            tmp_data = tmp_data.where(self.data.cnr > snr)
 
-        tmpData = tmpData.where(
+        tmp_data = tmp_data.where(
             (self.data.elevation != 90) & (self.data.azimuth == azm), drop=True
         )
 
-        return tmpData
+        return tmp_data
 
 
 # it removes the STE below cloud layer
-class secondTripEchoFilter:
+class SecondTripEchoFilter:
     """Boundary layer second trip echoes filter
 
     This filter minimises the presence of second
@@ -129,7 +129,7 @@ class secondTripEchoFilter:
         1 indicates cloud and 0 indicates no cloud.
         (THIS MAKS IS NOT NEEDED NOW)
 
-    nProf : int
+    n_prof : int
         number of profiles used to calculating the anomaly
 
     center : bool, optional
@@ -140,15 +140,15 @@ class secondTripEchoFilter:
         minimum number of profiles used for calculating the
         mean value
 
-    nStd : int
+    n_std : int
         Multiplication factor for defining the size of the
         window to keep the data. The filter removes any
-        anomaly larger than nStd * std
+        anomaly larger than n_std * std
 
-    strH : str
+    str_h : str
         starting hour for calculating the anomaly
 
-    endH : str
+    end_h : str
         end hour for calculating the anomaly
 
     Returns
[…]
