### 2022-12-10 3e7ae58
replacing configDic by config_dic
