### 2022-12-10 d75ff2c
replacing loadConfFile by load_conf_file
