+++ lidarSuit/dataAttributesL1.py
@@ -44,7 +44,7 @@ class loadAttributes:
 
         """
 
-        configInfo = configurations(lst=None).loadConfFile()
+        configInfo = configurations(lst=None).load_conf_file()
 
         tmpAtt = {
             "Conventions": "Cf/Radial 2.0",
