### 2022-12-10 bc6d77e
replacing configJS by config_js
