### 2022-12-10 d10140e
replacing generateConf by generate_conf
