+++ lidarSuit/lstConfig.py
@@ -21,16 +21,16 @@ class configurations:
 
     def __init__(self, lst=None):
 
-        self.loadVersion(lst)
-        self.loadReference()
-        self.loadInstitution()
-        self.loadInstrument()
-        self.loadSite()
-        self.loadContact()
-        self.loadEmail()
-        self.loadComments()
+        self.load_version(lst)
+        self.load_reference()
+        self.load_institution()
+        self.load_instrument()
+        self.load_site()
+        self.load_contact()
+        self.load_email()
+        self.load_comments()
 
-    def loadVersion(self, lst):
+    def load_version(self, lst):
         """
         It identifies the lidarSuit version
         and writes it to the configuration file
@@ -43,13 +43,13 @@ class configurations:
         """
 
         if lst == None:
-            self.lstVersion = "temporary config file"
+            self.lst_version = "temporary config file"
         else:
-            self.lstVersion = lst.__version__
+            self.lst_version = lst.__version__
 
         return self
 
-    def loadInstitution(self, institution="institution name"):
+    def load_institution(self, institution="institution name"):
         """
         It defines the institution affiliation name
 
@@ -64,7 +64,7 @@ class configurations:
 
         return self
 
-    def loadInstrument(self, instrument="instrument name"):
+    def load_instrument(self, instrument="instrument name"):
         """
         It defines the instrument name
 
@@ -79,7 +79,7 @@ class configurations:
 
         return self
 
-    def loadSite(self, site="site name"):
+    def load_site(self, site="site name"):
         """
         It defines the name of the experimental site
 
@@ -94,7 +94,7 @@ class configurations:
 
         return self
 
-    def loadContact(self, contact="contact person"):
+    def load_contact(self, contact="contact person"):
         """
         It defines the author's name
 
@@ -109,7 +109,7 @@ class configurations:
 
         return self
 
-    def loadEmail(self, email="contact email"):
+    def load_email(self, email="contact email"):
         """
         It defines the contacting email
 
@@ -123,7 +123,7 @@ class configurations:
 
         return self
 
-    def loadReference(self, reference="Generated by lidarSuit version: {0}"):
+    def load_reference(self, reference="Generated by lidarSuit version: {0}"):
         """
         It loads the lidarSuit's version used for
         processing the data
@@ -135,11 +135,11 @@ class configurations:
 
         """
 
-        self.references = reference.format(self.lstVersion)
+        self.references = reference.format(self.lst_version)
 
         return self
 
-    def loadComments(self, comments="General comments"):
+    def load_comments(self, comments="General comments"):
         """
         It defines additional comments
 
@@ -197,12 +197,12 @@ class configurations:
             self.generate_conf()
             config_dic = json.load(open(file_path))
 
-        self.loadReference(config_dic["references"])
-        self.loadInstitution(config_dic["institution"])
-        self.loadInstrument(config_dic["instrument_name"])
-        self.loadComments(config_dic["comments"])
-        self.loadSite(config_dic["site_name"])
-        self.loadContact(config_dic["contact_person"])
-        self.loadEmail(config_dic["email"])
+        self.load_reference(config_dic["references"])
+        self.load_institution(config_dic["institution"])
+        self.load_instrument(config_dic["instrument_name"])
+        self.load_comments(config_dic["comments"])
+        self.load_site(config_dic["site_name"])
+        self.load_contact(config_dic["contact_person"])
+        self.load_email(config_dic["email"])
 
         return self
