### 2022-12-10 5bb2b59
replacing lst == None by lst is None
