### 2022-12-10 2589a5d
Replacing camel case on lstConfig (#69)
* replacing configDic by config_dic

* replacing filePath by file_path

* replacing loadConfFile by load_conf_file

* replacing configFile by config_file

* replacing configJS by config_js

* replacing generateConf by generate_conf

* replacing camel case by snake case

* applying Pascal case to the class name

* replacing lst == None by lst is None

* replacing configurations by Configurations
