+++ lidarSuit/lstConfig.py
@@ -1,3 +1,4 @@
+
 """Module for configuring the data global attribute
 
 """
@@ -5,7 +6,7 @@
 import json
 
 
-class configurations:
+class Configurations:
     """Global attributes definition
 
     This class defines all global attributes
@@ -20,16 +21,16 @@ class configurations:
 
     def __init__(self, lst=None):
 
-        self.loadVersion(lst)
-        self.loadReference()
-        self.loadInstitution()
-        self.loadInstrument()
-        self.loadSite()
-        self.loadContact()
-        self.loadEmail()
-        self.loadComments()
+        self.load_version(lst)
+        self.load_reference()
+        self.load_institution()
+        self.load_instrument()
+        self.load_site()
+        self.load_contact()
+        self.load_email()
+        self.load_comments()
 
-    def loadVersion(self, lst):
+    def load_version(self, lst):
         """
         It identifies the lidarSuit version
         and writes it to the configuration file
@@ -41,14 +42,14 @@ class configurations:
 
         """
 
-        if lst == None:
-            self.lstVersion = "temporary config file"
+        if lst is None:
+            self.lst_version = "temporary config file"
         else:
-            self.lstVersion = lst.__version__
+            self.lst_version = lst.__version__
 
         return self
 
-    def loadInstitution(self, institution="institution name"):
+    def load_institution(self, institution="institution name"):
         """
         It defines the institution affiliation name
 
@@ -63,7 +64,7 @@ class configurations:
 
         return self
 
-    def loadInstrument(self, instrument="instrument name"):
+    def load_instrument(self, instrument="instrument name"):
         """
         It defines the instrument name
 
@@ -78,7 +79,7 @@ class configurations:
 
         return self
 
-    def loadSite(self, site="site name"):
+    def load_site(self, site="site name"):
         """
         It defines the name of the experimental site
 
@@ -93,7 +94,7 @@ class configurations:
 
         return self
 
-    def loadContact(self, contact="contact person"):
+    def load_contact(self, contact="contact person"):
         """
         It defines the author's name
 
@@ -108,7 +109,7 @@ class configurations:
 
         return self
 
-    def loadEmail(self, email="contact email"):
+    def load_email(self, email="contact email"):
         """
         It defines the contacting email
 
@@ -122,7 +123,7 @@ class configurations:
 
         return self
 
-    def loadReference(self, reference="Generated by lidarSuit version: {0}"):
+    def load_reference(self, reference="Generated by lidarSuit version: {0}"):
         """
         It loads the lidarSuit's version used for
         processing the data
@@ -134,11 +135,11 @@ class configurations:
 
         """
 
-        self.references = reference.format(self.lstVersion)
+        self.references = reference.format(self.lst_version)
 
         return self
 
-    def loadComments(self, comments="General comments"):
+    def load_comments(self, comments="General comments"):
         """
         It defines additional comments
 
@@ -152,56 +153,56 @@ class configurations:
 
         return self
 
-    def generateConf(self):
+    def generate_conf(self):
         """
         It writes and saves all defined global attributes.
 
         """
 
-        configDic = {}
+        config_dic = {}
 
-        configDic["references"] = self.references
-        configDic["institution"] = self.institution
-        configDic["instrument_name"] = self.instrument
-        configDic["site_name"] = self.site
-        configDic["comments"] = self.comments
-        configDic["contact_person"] = self.contact
-        configDic["email"] = self.email
+        config_dic["references"] = self.references
+        config_dic["institution"] = self.institution
+        config_dic["instrument_name"] = self.instrument
+        config_dic["site_name"] = self.site
+        config_dic["comments"] = self.comments
+        config_dic["contact_person"] = self.contact
+        config_dic["email"] = self.email
 
[…]
