### 2022-12-10 05c231a
Applying Pascal case to the class name
