### 2022-12-10 eb57287
replacing attrsDic by attrs_dic
