### 2022-12-10 15445a7
replacing writeGlobalAttrs by write_global_attrs
