### 2022-12-10 36da981
replacing configInfo by configInfo
