### 2022-12-10 1bf4c14
replacing variablesAttrs by variables_attrs
