### 2022-12-10 1a041f3
replacing writeCoordsAttrs by write_coords_attrs
