### 2022-12-10 67b8b14
replacing writeVariablesAttrs by write_variables_attrs
