### 2022-12-10 aea33e6
defining exception type
