### 2022-12-10 d51a0f6
Refactor/camel case data attrs (#70)
* Applying Pascal case to the class name

* Adjusting line lenght

* replacing attrsDic by attrs_dic

* replacing writeGlobalAttrs by write_global_attrs

* replacing configInfo by configInfo

* replacing tmpAtt by tmp_att

* replacing variablesAttrs by variables_attrs

* replacing writeCoordsAttrs by write_coords_attrs

* replacing writeVariablesAttrs by write_variables_attrs

* defining exception type
