### 2022-12-10 f498f0d
Replacing getLidarData by GetLidarData
