### 2022-12-10 d92ded0
Removing useless return
