+++ lidarSuit/dataOperator.py
@@ -83,7 +83,7 @@ class dataOperations:
         for filePath in self.dataPaths:
 
             try:
-                tmpFile = GetLidarData(filePath).OpenLidarFile()
+                tmpFile = GetLidarData(filePath).open_lidar_file()
                 self.logger.debug(f"reading file: {filePath}")
             except:
                 self.logger.warning(f"This file has a problem: {filePath}")
@@ -722,7 +722,7 @@ class dbsOperations:
         for file in file_list:
 
             try:
-                fileToMerge = GetLidarData(file).OpenLidarFile()
+                fileToMerge = GetLidarData(file).open_lidar_file()
                 self.logger.debug(f"reading file: {file}")
             except:
                 self.logger.warning(f"This file has a problem: {file}")
