### 2022-12-10 1ade716
Replacing camel case on lidar_code  (#71)
* Replacing getLidarData by GetLidarData

* Replacing fileName by file_name

* Removing useless return

* Replacing openLidarFile by OpenLidarFile

* Removing not used module

* Replacing OpenLidarFile by open_lidar_file
