+++ lidarSuit/windPropRetrieval6Beam.py
@@ -4,7 +4,6 @@
 
 import numpy as np
 import xarray as xr
-import pandas as pd
 
 from .utilities import util
 
