### 2022-12-12 8e1434d
Replacing varComp by var_comp
