### 2022-12-12 66e88b2
Replacing getVarianceDS by get_variance_ds
