### 2022-12-12 9edb97e
Replacing sixBeamMethod by six_beam_method
