### 2022-12-12 d97f554
Replacing git sixBeamMethod by SixBeamMethod
