### 2022-12-12 18a1ee7
Replacing get_SIGMA by get_sigma
