### 2022-12-12 d4edaa9
Replacing get_S by get_s_matrix and S by s_matrix
