### 2022-12-12 bfffc2c
Replacing getVariance by get_variance
