### 2022-12-12 4028606
Replacing get_M by get_m_matrix
