### 2022-12-12 c88b50a
Replacing M by m_matrix
