### 2022-12-12 b6fa010
Replacing SIGMA by sigma_matrix
