### 2022-12-12 825a090
Refactor/camel case 6beam (#72)
* Removing non used module

* Replacing varCompDS by var_comp_ds

* replacing tmpData by tmp_data

* Correcting docstring return

* Replacing varCompName by var_comp_name

* Replacing varComp by var_comp

* Replacing getVarianceDS by get_variance_ds

* Replacing sixBeamMethod by six_beam_method

* Replacing git sixBeamMethod by SixBeamMethod

* Replacing rVariances by radial_variances

* Replacing get_SIGMA by get_sigma

* Replacing get_S by get_s_matrix and S by s_matrix

* Replacing getVariance by get_variance

* Replacing interpDataTransf by interp_data_trans
[…]
