+++ lidarSuit/windPropRetrieval6Beam.py
@@ -4,12 +4,11 @@
 
 import numpy as np
 import xarray as xr
-import pandas as pd
 
 from .utilities import util
 
 
-class sixBeamMethod:
+class SixBeamMethod:
 
     """6 beam method
 
@@ -35,7 +34,7 @@ class sixBeamMethod:
 
     Returns
     -------
-    varCompDS : xarray.DataSet
+    var_comp_ds : xarray.DataSet
         a dataset of the eynolds stress tensor
         matrix elementes
 
@@ -49,16 +48,16 @@ class sixBeamMethod:
         # self.timeFreq = freq
         # self.timeFreq = freq
 
-        self.get_M()
-        self.get_M_inv()
-        self.rVariances = {}
-        self.calcVariances(data, freq, freq90)
+        self.get_m_matrix()
+        self.get_m_matrix_inv()
+        self.radial_variances = {}
+        self.calc_variances(data, freq, freq90)
 
-        self.get_S()
-        self.get_SIGMA()
-        self.getVarianceDS()
+        self.get_s_matrix()
+        self.get_sigma()
+        self.get_variance_ds()
 
-    def get_M(self):
+    def get_m_matrix(self):
 
         """
         This method populates the coefficient matrix (M).
@@ -75,16 +74,16 @@ class sixBeamMethod:
         """
 
         phis = np.append(np.ones_like(self.azm) * self.elv, np.array([90]))
-        phisRad = np.deg2rad(phis)
+        phis_rad = np.deg2rad(phis)
 
         thetas = np.append(self.azm, np.array([0]))
-        thetasRad = np.deg2rad(thetas)
+        thetas_rad = np.deg2rad(thetas)
 
-        M = np.ones((len(phis), len(thetas))) * np.nan
+        m_matrix = np.ones((len(phis), len(thetas))) * np.nan
 
-        for i, theta in enumerate(thetasRad):
+        for i, theta in enumerate(thetas_rad):
 
-            phi = phisRad[i]
+            phi = phis_rad[i]
 
             ci1 = np.cos(phi) ** 2 * np.sin(theta) ** 2
             ci2 = np.cos(phi) ** 2 * np.cos(theta) ** 2
@@ -95,40 +94,40 @@ class sixBeamMethod:
             ci5 = np.cos(phi) * np.sin(phi) * np.sin(theta)
             ci6 = np.cos(phi) * np.sin(phi) * np.cos(theta)
 
-            Mline = np.array([ci1, ci2, ci3, ci4 * 2, ci5 * 2, ci6 * 2])
+            m_matrix_line = np.array([ci1, ci2, ci3, ci4 * 2, ci5 * 2, ci6 * 2])
 
-            M[i] = Mline
+            m_matrix[i] = m_matrix_line
 
-        self.M = M
+        self.m_matrix = m_matrix
 
         return self
 
-    def get_M_inv(self):
+    def get_m_matrix_inv(self):
 
         """
         This method calculates the inverse matrix of M.
         """
 
-        self.M_inv = np.linalg.inv(self.M)
+        self.m_matrix_inv = np.linalg.inv(self.m_matrix)
 
         return self
 
     ### new approach to calculate the variances ##############
     #
     #
-    def calcVariances(self, data, freq, freq90):
+    def calc_variances(self, data, freq, freq90):
 
-        interpDataTransf = data.dataTransf.interp(
+        interp_data_transf = data.dataTransf.interp(
             time=data.dataTransf90.time, method="nearest"
         )
-        self.getVariance(interpDataTransf, freq=freq)
-        self.getVariance(
+        self.get_variance(interp_data_transf, freq=freq)
+        self.get_variance(
             -1 * data.dataTransf90, freq=freq90, name="rVariance90"
[…]
