+++ lidarSuit/utilities.py
@@ -1,7 +1,8 @@
 import os
-import gdown
 import shutil
 import glob
+
+import gdown
 import pandas as pd
 import numpy as np
 import xarray as xr
