### 2022-12-12 d145a1e
Replacing auxCloudMask by aux_cloud_mask
