### 2022-12-12 c3b5821
Replacing maskType by mask_type
