### 2022-12-12 cc06f6f
Replacing getTimeBins by get_time_bins
