+++ lidarSuit/utilities.py
@@ -14,17 +14,17 @@ class Util:
     This class contains useful tools
     """
 
-    def get_time_bins(selDay, freq="10min"):
+    def get_time_bins(sel_day, freq="10min"):
         """Bins estimation
 
         Creating time bins for a given day and time resolution
 
         """
 
-        start = selDay.strftime("%Y%m%d")
+        start = sel_day.strftime("%Y%m%d")
         start_time = pd.to_datetime(f"{start} 00:00:00")
 
-        end = (selDay + pd.to_timedelta(1, "D")).strftime("%Y%m%d")
+        end = (sel_day + pd.to_timedelta(1, "D")).strftime("%Y%m%d")
         end_time = pd.to_datetime(f"{end} 00:00:00")
 
         timeBins = pd.date_range(start_time, end_time, freq=freq)
