### 2022-12-12 f39b13d
Replacing cloudMask by cloud_mask
