### 2022-12-12 fa649bb
Replacing cleanCeiloData by clean_ceilo_data
