### 2022-12-12 a59550a
Replacing cleanRadarData by clean_radar_data
