### 2022-12-12 386fbd0
Replacing cleanCeilo by clean_ceilo
