### 2022-12-12 f9d3cc4
Replacing positiveBeta by positive_beta
