### 2022-12-12 114fb16
Replacing cleanRadar by clean_radar
