### 2022-12-12 450e755
Replacing getCloudMask2D by get_cloud_mask_2d
