### 2022-12-12 9dcea09
Replacing getTimeMask by get_time_mask
