### 2022-12-12 27fe146
Replacing camel case from utilities module (#73)
* Changing import order

* Replacing timeCloudMask by time_cloud_mask

* Replacing highCloudLayer by high_cloud_layer

* Replacing auxCloudMask by aux_cloud_mask

* Replacing maskType by mask_type

* Replacing util by Util

* Replacing getTimeBins by get_time_bins

* Raplacing selDay by sel_day

* Replacing timeBins by time_bins

* Replacing = by is

* Replacing cloudMask by cloud_mask

* Applying Pascal case to the class name

* Replacing ceiloData by ceilo_data

* Replacing radarData by radar_data

* Replacing wcData by wc_data

* Replacing cl
[…]
