+++ lidarSuit/utilities.py
@@ -1,34 +1,35 @@
 import os
-import gdown
 import shutil
 import glob
+
+import gdown
 import pandas as pd
 import numpy as np
 import xarray as xr
 
 
-class util:
+class Util:
 
     """
     This class contains useful tools
     """
 
-    def getTimeBins(selDay, freq="10min"):
+    def get_time_bins(sel_day, freq="10min"):
         """Bins estimation
 
         Creating time bins for a given day and time resolution
 
         """
 
-        start = selDay.strftime("%Y%m%d")
+        start = sel_day.strftime("%Y%m%d")
         start_time = pd.to_datetime(f"{start} 00:00:00")
 
-        end = (selDay + pd.to_timedelta(1, "D")).strftime("%Y%m%d")
+        end = (sel_day + pd.to_timedelta(1, "D")).strftime("%Y%m%d")
         end_time = pd.to_datetime(f"{end} 00:00:00")
 
-        timeBins = pd.date_range(start_time, end_time, freq=freq)
+        time_bins = pd.date_range(start_time, end_time, freq=freq)
 
-        return timeBins
+        return time_bins
 
     def get_sample_data(sample_path, file_type):
         """Downloading data
@@ -66,25 +67,25 @@ class util:
             if os.path.isdir(f"{sample_path}{file_type}/"):
                 file_list = sorted(glob.glob(f"{sample_path}{file_type}/*.nc"))
 
-                if bool(file_list) == False:
-                    util.get_sample_data(sample_path, file_type)
+                if bool(file_list) is False:
+                    Util.get_sample_data(sample_path, file_type)
                     file_list = sorted(
                         glob.glob(f"{sample_path}{file_type}/*.nc")
                     )
 
             else:
-                util.get_sample_data(sample_path, file_type)
+                Util.get_sample_data(sample_path, file_type)
                 file_list = sorted(glob.glob(f"{sample_path}{file_type}/*.nc"))
 
         else:
             os.makedirs(sample_path)
-            util.get_sample_data(sample_path, file_type)
+            Util.get_sample_data(sample_path, file_type)
             file_list = sorted(glob.glob(f"{sample_path}{file_type}/*.nc"))
 
         return file_list
 
 
-class cloudMask:
+class CloudMask:
 
     """
     This class generates the time-height cloud mask
@@ -92,96 +93,96 @@ class cloudMask:
     from lidar and ceilometer.
     """
 
-    def __init__(self, wcData=None, ceiloData=None, radarData=None):
+    def __init__(self, wc_data=None, ceilo_data=None, radar_data=None):
 
-        self.ceiloData = ceiloData
-        self.radarData = radarData
-        self.wcData = wcData
+        self.ceilo_data = ceilo_data
+        self.radar_data = radar_data
+        self.wc_data = wc_data
 
-        self.callMethods()
+        self.call_methods()
 
-    def callMethods(self):
+    def call_methods(self):
 
-        if self.ceiloData is None or self.radarData is None:
+        if self.ceilo_data is None or self.radar_data is None:
 
-            self.getTimeMask(maskType="aux")
+            self.get_time_mask(mask_type="aux")
 
         else:
 
-            self.cleanCeilo()
-            self.cleanRadar()
-            self.getCloudMask2D()
-            self.getTimeMask(maskType="real")
+            self.clean_ceilo()
+            self.clean_radar()
+            self.get_cloud_mask_2d()
+            self.get_time_mask(mask_type="real")
 
-    def cleanCeilo(self):
+    def clean_ceilo(self):
 
-        positiveBeta = self.ceiloData.beta_raw.where(
-            self.ceiloData.beta_raw > 0
[…]
