### 2022-12-13 f7b5d50
Applying Pascal case to  plotSettings
