### 2022-12-13 84e64e2
Replacing updateSettings by update_settings
