### 2022-12-13 2feb8cb
Replacing plotSetup by plot_setup
