### 2022-12-13 8bae98b
Replacing viewOrigVar by view_orig_var
