### 2022-12-13 b5c4c53
Replacing figPath by fig_path
