### 2022-12-13 ec56f5f
Replacing minTime by min_time
