### 2022-12-13 57c3b52
Replacing strName by string_name
