### 2022-12-13 f37332d
Replacing viewRetVar by view_ret_var
