### 2022-12-13 51ffa1f
Replacing plotDataAZM by plot_data_azm
