### 2022-12-13 12933bc
Replacing camel case from visualization module (#74)
* Applying Pascal case to  plotSettings

* Replacing updateSettings by update_settings

* Replacing fs by font_size

* Replacing plotSetup by plot_setup

* Applying Pascal case to visualizer

* Replacing viewOrigVar by view_orig_var

* Replacing varName by var_name

* Replacing plotID by plot_id

* Replacing figPath by fig_path

* Replacing namePrefix by name_prefix

* Replacing minTime by min_time

* Replacing maxTime by max_time

* Replacing strName by string_name

* Replacing viewRetVar by view_ret_var

* Replacing plotData by plot_data


[…]
