+++ lidarSuit/visualization.py
@@ -7,16 +7,16 @@ import matplotlib.dates as mdates
 from .filters import Filtering
 
 
-class plotSettings:
+class PlotSettings:
     def __init__(self, mpl, style="dark_background"):
 
         self.mpl = mpl
         self.style = style
-        # self.updateSettings()
+        # self.update_settings()
 
-    def updateSettings(self):
+    def update_settings(self):
 
-        fs = 16
+        font_size = 16
 
         # mpl.style.use('seaborn')
         self.mpl.style.use(self.style)
@@ -24,14 +24,14 @@ class plotSettings:
         self.mpl.rcParams["figure.dpi"] = 80
         self.mpl.rcParams["savefig.dpi"] = 100
 
-        self.mpl.rcParams["font.size"] = fs
-        self.mpl.rcParams["legend.fontsize"] = fs
-        self.mpl.rcParams["figure.titlesize"] = fs
+        self.mpl.rcParams["font.size"] = font_size
+        self.mpl.rcParams["legend.fontsize"] = font_size
+        self.mpl.rcParams["figure.titlesize"] = font_size
 
-        self.mpl.rcParams["ytick.labelsize"] = fs
-        self.mpl.rcParams["xtick.labelsize"] = fs
-        self.mpl.rcParams["axes.titlesize"] = fs
-        self.mpl.rcParams["axes.labelsize"] = fs
+        self.mpl.rcParams["ytick.labelsize"] = font_size
+        self.mpl.rcParams["xtick.labelsize"] = font_size
+        self.mpl.rcParams["axes.titlesize"] = font_size
+        self.mpl.rcParams["axes.labelsize"] = font_size
 
         self.mpl.rcParams["legend.fancybox"] = True
         self.mpl.rcParams["legend.framealpha"] = 0.7
@@ -42,7 +42,7 @@ class plotSettings:
 
         return self
 
-    def plotSetup(plot):
+    def plot_setup(plot):
 
         plt.setp(plot.axes.xaxis.get_majorticklabels(), rotation=0)
         locator = mdates.AutoDateLocator()
@@ -54,56 +54,56 @@ class plotSettings:
         return plot
 
 
-class visualizer:
+class Visualizer:
     def __init__(self, data):
 
         self.data = data
 
-    def viewOrigVar(
+    def view_orig_var(
         self,
-        varName,
+        var_name,
         cmap="Spectral",
         vmin=-1,
         vmax=1,
         elv="90",
         azm="-",
         save=False,
-        plotID=None,
-        figPath=None,
-        namePrefix=None,
+        plot_id=None,
+        fig_path=None,
+        name_prefix=None,
         show=False,
-        minTime=None,
-        maxTime=None,
+        min_time=None,
+        max_time=None,
     ):
 
-        if plotID == "rad_wind_speed_panel":
+        if plot_id == "rad_wind_speed_panel":
 
             tmp_data = self.data
-            self.plotDataAZM(
-                dataNon90=tmp_data,
+            self.plot_data_azm(
+                data_non_90=tmp_data,
                 cmap=cmap,
                 vmin=vmin,
                 vmax=vmax,
-                plotID=plotID,
-                figPath=figPath,
+                plot_id=plot_id,
+                fig_path=fig_path,
                 save=save,
                 show=show,
-                minTime=minTime,
-                maxTime=maxTime,
+                min_time=min_time,
+                max_time=max_time,
             )
 
         else:
-            tmp_data = Filtering(self.data).get_vertical_obs_comp(varName)
+            tmp_data = Filtering(self.data).get_vertical_obs_comp(var_name)
 
-            if namePrefix:
-                strName = "{}_{}".format(
-                    namePrefix, tmp_data.attrs["standard_name"]
[…]
