### 2022-12-14 3ec53ab
Replacing tmpBeta by tmp_beta
