### 2022-12-14 3928595
Replacing getBeta by get_beta
