### 2022-12-14 f4fd1e2
Replacing tmpWindW by tmp_wind_w
