### 2022-12-14 759beec
Replacing retVertWindData by ret_vert_wind_data
