### 2022-12-14 a9de593
Replacing tmpWindProp by tmp_wind_prop
