### 2022-12-14 0dc9684
Replacing retrieveWind by retrieve_wind
