### 2022-12-14 1843d8d
Replacing calcHorWindDir by calc_hor_wind_dir
