### 2022-12-14 5fa02e8
Replacing calcHorWindSpeed by calc_hor_wind_speed
