### 2022-12-14 4b66cc3
Replacing compUW by comp_uw
