### 2022-12-14 b982a29
Replacing compVS by comp_vs
