### 2022-12-14 d1c933e
Replacing compWindSpeed by comp_wind_speed
