### 2022-12-14 6f82a92
Replacing calcHorWindComp_continuous by calc_hor_wind_comp_continuous
