### 2022-12-15 7d08bbf
Replacing meanTimeUE by mean_time_ue
