### 2022-12-15 d978383
Replacing fftWindPropRet by FourierTransfWindMethod
