### 2022-12-15 e0e8192
Replacing compVN by comp_vn
