### 2022-12-15 d95e42a
Replacing meanTimeVN by mean_time_vn
