### 2022-12-15 036e8e8
Replacing calcHorWindComp_single_dbs by calc_hor_wind_comp_single_dbs
