### 2022-12-15 c897869
Replacing verWindSpeed by ver_wind_speed
