### 2022-12-15 134300c
Replacing correctVertWindComp by correct_vert_wind_comp
