### 2022-12-15 54a10a6
Replacing statusFilter by status_filter
