### 2022-12-15 22c9731
Replacing rangeVal90 by range_val_90
