### 2022-12-15 68810ff
Replacing rangeValNon90 by range_val_non_90
