### 2022-12-15 0ffc785
Replacing getWindProperties5Beam by GetWindProperties5Beam
