### 2022-12-15 09687df
Replacing getWindConpV by get_wind_comp_v
