### 2022-12-15 e074bc3
Replacing getWindConpU by get_wind_comp_u
