### 2022-12-15 430b06f
Replacing get_wind_conp_v by get_wind_comp_v
