### 2022-12-15 c5f6f97
Replacing getHorWindSpeed by get_hor_wind_speed
