### 2022-12-15 529922d
Replacing getAzmWind by get_azm_wind
