### 2022-12-15 e328f78
Replacing getRadWindSpeed by get_rad_wind_speed
