### 2022-12-15 d4a8beb
Replacing getWindDir by get_wind_dir
