### 2022-12-15 ead359e
Replacing getPhase by get_phase
