### 2022-12-15 8052822
Replacing getCompAmp by get_comp_amp
