### 2022-12-15 1afc8bf
Replacing radWindSpeed by rad_wind_speed
