### 2022-12-15 e3b465c
Replacing camel case from windPropRetrieval module (#75)
* Replacing loadAttrs by load_attrs

* Replacing tmpBeta by tmp_beta

* Replacing getBeta by get_beta

* Replacing tmpWindW by tmp_wind_w

* Replacing retVertWindData by ret_vert_wind_data

* Replacing tmpWindProp by tmp_wind_prop

* Replacing retHorWindData by ret_hor_wind_data

* Replacing transfdData by transfd_data

* Replacing retrieveWind by retrieve_wind

* Applying Pascal case to class name

* Replacing windDir by wind_dir

* Replacing calcHorWindDir by calc_hor_wind_dir

* Replacing horWindSpeed by hor_wind_speed

* Replacing ca
[…]
