+++ lidarSuit/windPropRetrieval.py
@@ -12,7 +12,7 @@ module_logger = logging.getLogger("lidarSuit.windPropRetrieval")
 module_logger.debug("loading windPropRetrieval")
 
 
-class fftWindPropRet:
+class FourierTransfWindMethod:
     """FFT wind retrieval method
 
     It is a fft wind retrieval method. It was proposed
@@ -21,42 +21,42 @@ class fftWindPropRet:
 
     Parameters
     ----------
-    dopplerObs : xarray.DataArray
+    doppler_obs : xarray.DataArray
         It should be a DataArray of slanted Doppler velocity
         observations as function of the azimuthal angle.
         It must have a coordinate called azm.
 
     Returns
     -------
-    windProp : xarray.Dataset
+    wind_prop : xarray.Dataset
         A Dataset containing the horizontal wind speed
         and direction. It also includes the zonal and
         meridional wind components
 
     """
 
-    def __init__(self, dopplerObs: xr.DataArray):
+    def __init__(self, doppler_obs: xr.DataArray):
 
         self.logger = logging.getLogger(
-            "lidarSuit.windPropRetrieval.fftWindPropRet"
+            "lidarSuit.windPropRetrieval.FourierTransfWindMethod"
         )
-        self.logger.info("creating an instance of fftWindPropRet")
+        self.logger.info("creating an instance of FourierTransfWindMethod")
 
-        if not isinstance(dopplerObs, xr.DataArray):
+        if not isinstance(doppler_obs, xr.DataArray):
             self.logger.error("wrong data type: expecting a xr.DataArray")
             raise TypeError
 
-        self.dopplerObs = dopplerObs
+        self.doppler_obs = doppler_obs
         #         self.elv = elv
-        self.getCompAmp()
-        self.getPhase()
-        self.getRadWindSpeed()
-        self.getHorWindSpeed()
-        self.getWindDir()
-        self.getWindConpU()
-        self.getWindConpV()
+        self.get_comp_amp()
+        self.get_phase()
+        self.get_rad_wind_speed()
+        self.get_hor_wind_speed()
+        self.get_wind_dir()
+        self.get_wind_comp_u()
+        self.get_wind_comp_v()
 
-    def getCompAmp(self):
+    def get_comp_amp(self):
         """First harmonic amplitude
 
         It calculates the complex amplitudes from the
@@ -66,32 +66,32 @@ class fftWindPropRet:
 
         self.logger.info("calculating the complex amplitude")
 
-        self.compAmp = xrft.fft(self.dopplerObs, dim=["azm"]).isel(freq_azm=-2)
+        self.comp_amp = xrft.fft(self.doppler_obs, dim=["azm"]).isel(freq_azm=-2)
 
         return self
 
-    def getPhase(self):
+    def get_phase(self):
         """First harmonic phase
 
         It calculates the phase of the first harmonic from the complex
-        amplitude used for retrieving wind direction, i.e. compAmp.
+        amplitude used for retrieving wind direction, i.e. comp_amp.
         """
 
         self.logger.info("calculating the phase from the complex amplitude")
 
         self.phase = -np.rad2deg(
-            np.arctan2(self.compAmp.imag, self.compAmp.real)
+            np.arctan2(self.comp_amp.imag, self.comp_amp.real)
         )
 
         self.phase.attrs = {
             "long_name": "Retrived phase",
             "units": "degree",
-            "comments": "Phase derived from compAmp variable using the FFT method.",
+            "comments": "Phase derived from comp_amp variable using the FFT method.",
         }
 
         return self
 
-    def getWindDir(self):
[…]
