### 2022-12-16 54c5a7b
Replacing toDataArray by convert_to_data_array
