### 2022-12-16 8f823f4
Replacing deltaGridMin by delta_grid_min
