### 2022-12-16 d12b2ee
Replacing getNearestIndexM2 by get_nearest_index_method_2
