### 2022-12-16 b9add4d
Replacing origGrid by orig_grid
