### 2022-12-16 9d4ac7b
Replacing origGrid by orig_grid
