### 2022-12-16 6d528c5
Replacing calcDeltaGrid by calc_delt_grid
