### 2022-12-16 1bf09c6
Replacing getTimeRef by get_time_ref
