### 2022-12-16 ec7cf42
Replacing timeRefSec by time_ref_sec
