### 2022-12-16 6f6ab43
Replacing xrDataArray by xr_data_array
