### 2022-12-16 c182439
Replacing getResampledData by get_resampled_data
