### 2022-12-16 d058460
Replacing newRange by new_range
