### 2022-12-16 633bf7a
Replacing tmpRadWind by tmp_rad_wind
