### 2022-12-16 ffdd961
Replacing dopWindArr by dop_wind_arr
