### 2022-12-16 4923095
Replacing getCoordNon90 by get_coord_non_90
