+++ lidarSuit/windPropRetrieval6Beam.py
@@ -118,11 +118,11 @@ class SixBeamMethod:
     def calc_variances(self, data, freq, freq90):
 
         interp_data_transf = data.dataTransf.interp(
-            time=data.dataTransf90.time, method="nearest"
+            time=data.data_transf_90.time, method="nearest"
         )
         self.get_variance(interp_data_transf, freq=freq)
         self.get_variance(
-            -1 * data.dataTransf90, freq=freq90, name="rVariance90"
+            -1 * data.data_transf_90, freq=freq90, name="rVariance90"
         )  # think about the -1 coefficient
 
         return self
