### 2022-12-16 43d9ed7
Replacing getRestructuredData by GetRestructuredData
