+++ lidarSuit/windPropRetrieval6Beam.py
@@ -22,7 +22,7 @@ class SixBeamMethod:
     ----------
     data : object
         an instance of the object generated by the
-        lst.getRestructuredData()
+        lst.GetRestructuredData()
 
     freq : int
         number of profiles used to calculate
