### 2022-12-16 c412578
Replacing tmpMerged by tmp_merged
