### 2022-12-16 c56c153
Replacing mergeDataM2 by merge_data_method_2
