### 2022-12-16 d923577
Replacing mergeDataM1 by merge_data_method_1
