### 2022-12-16 fc8e366
Replacing getMergeData by get_merge_data
