### 2022-12-16 3195aee
Replacing mergedData by merged_data
