### 2022-12-16 35bc7de
Replacing camel case from dataOperator module (#76)
* Replacing fileToMerge by file_to_merge

* Replacing merge2DS by merge_2_ds

* Replacing meanTimeDA by mean_time_da

* Replacing meanTimeNS by mean_time_ns

* Replacing meanTime by mean_time

* Replacing lidarDS by lidar_ds

* Replacing mergeData by merge_data

* Replacing varList by var_list

* Replacing fileList by file_list

* Replacing mergedDS by merged_ds

* Replacing dbsOperations by DbsOperations

* Replacing tmpDT by tmp_dt

* Replacing toDataArray by convert_to_data_array

* Replacing timeIndex by time_index

* Replacing t by posit
[…]
