+++ lidarSuit/windPropRetrieval6Beam.py
@@ -22,7 +22,7 @@ class SixBeamMethod:
     ----------
     data : object
         an instance of the object generated by the
-        lst.getRestructuredData()
+        lst.GetRestructuredData()
 
     freq : int
         number of profiles used to calculate
@@ -43,10 +43,10 @@ class SixBeamMethod:
     def __init__(self, data, freq=10, freq90=10):
 
         #         self.elv = 45
-        self.elv = data.dataTransf.elv.values
-        self.azm = data.dataTransf.azm.values
-        # self.timeFreq = freq
-        # self.timeFreq = freq
+        self.elv = data.data_transf.elv.values
+        self.azm = data.data_transf.azm.values
+        # self.time_freq = freq
+        # self.time_freq = freq
 
         self.get_m_matrix()
         self.get_m_matrix_inv()
@@ -117,12 +117,12 @@ class SixBeamMethod:
     #
     def calc_variances(self, data, freq, freq90):
 
-        interp_data_transf = data.dataTransf.interp(
-            time=data.dataTransf90.time, method="nearest"
+        interp_data_transf = data.data_transf.interp(
+            time=data.data_transf_90.time, method="nearest"
         )
         self.get_variance(interp_data_transf, freq=freq)
         self.get_variance(
-            -1 * data.dataTransf90, freq=freq90, name="rVariance90"
+            -1 * data.data_transf_90, freq=freq90, name="rVariance90"
         )  # think about the -1 coefficient
 
         return self
@@ -167,7 +167,7 @@ class SixBeamMethod:
     #         The default size of this window is 10 minutes.
     #         """
 
-    #         time_bins = Util.get_time_bins(pd.to_datetime(data.time.values[0]), freq=self.timeFreq)
+    #         time_bins = Util.get_time_bins(pd.to_datetime(data.time.values[0]), freq=self.time_freq)
     #         groupedData = data.groupby_bins('time', time_bins)
 
     #         self.radial_variances[name] = groupedData.var(dim='time')#.apply(calcGroupVar)
