+++ lidarSuit/dataOperator.py
@@ -20,7 +20,7 @@ module_logger = logging.getLogger("lidarSuit.dataOperator")
 module_logger.debug("loading dataOperator")
 
 
-class dataOperations:
+class DataOperations:
 
     """Basic data manager
 
@@ -34,45 +34,45 @@ class dataOperations:
 
     Examples
     --------
-    >>> mergedDS = lidarSuit.dataOperations(fileList).mergedData
-    >>> mergedDS.to_netcdf(outputFilePath)
+    >>> merged_ds = lidarSuit.DataOperations(file_list).merged_data
+    >>> merged_ds.to_netcdf(output_file_path)
 
     Parameters
     ----------
-    dataPaths : list
+    data_paths : list
        List of paths of the original WindCube's output.
 
     Returns
     -------
     object : object
         it returns an object containing an instance of the
-        original files merged (.mergedData)
+        original files merged (.merged_data)
 
     """
 
-    def __init__(self, dataPaths, verbose=False):
+    def __init__(self, data_paths, verbose=False):
 
         self.logger = logging.getLogger(
-            "lidarSuit.dataOperator.dataOperations"
+            "lidarSuit.dataOperator.DataOperations"
         )
-        self.logger.info("creating an instance of dataOperations")
+        self.logger.info("creating an instance of DataOperations")
 
-        if bool(dataPaths) == False:
+        if bool(data_paths) is False:
             self.logger.error(
                 "lidarSuit stopped due to an empty list of files."
             )
             raise FileNotFoundError
 
         self.verbose = verbose
-        self.dataPaths = dataPaths
+        self.data_paths = data_paths
         self.tmp90 = xr.Dataset()
-        self.tmpNon90 = xr.Dataset()
+        self.tmp_non_90 = xr.Dataset()
 
-        self.elevationFilter()
-        self.renameVar90()
-        self.getMergeData()
+        self.elevation_filter()
+        self.rename_var_90()
+        self.get_merge_data()
 
-    def elevationFilter(self):
+    def elevation_filter(self):
         """
         It groups the data from the vertical and slanted observations
         and rounds the azimuth coordinate
@@ -80,39 +80,39 @@ class dataOperations:
 
         self.logger.info("coverting azimuth: from 360 to 0 degrees")
 
-        for filePath in self.dataPaths:
+        for file_path in self.data_paths:
 
             try:
-                tmpFile = GetLidarData(filePath).open_lidar_file()
-                self.logger.debug(f"reading file: {filePath}")
+                tmp_file = GetLidarData(file_path).open_lidar_file()
+                self.logger.debug(f"reading file: {file_path}")
             except:
-                self.logger.warning(f"This file has a problem: {filePath}")
+                self.logger.warning(f"This file has a problem: {file_path}")
 
             try:
-                elevation = tmpFile.elevation.round(1)
-                tmpFile["elevation"] = elevation
-                tmpFile["azimuth"] = tmpFile.azimuth.round(1)
-                tmpFile["azimuth"][tmpFile.azimuth == 360] = 0
+                elevation = tmp_file.elevation.round(1)
+                tmp_file["elevation"] = elevation
+                tmp_file["azimuth"] = tmp_file.azimuth.round(1)
+                tmp_file["azimuth"][tmp_file.azimuth == 360] = 0
             except:
-                self.logger.info(f"Problems reading elv and axm: {filePath}")
+                self.logger.info(f"Problems reading elv and axm: {file_path}")
 
             try:
                 self.tmp90 = xr.merge(
-                    [self.tmp90, tmpFile.where(elevation == 90, drop=True)]
[…]
