+++ tests/test_dataOperator.py
@@ -5,37 +5,37 @@ import xarray as xr
 import lidarSuit as lst
 
 
-def test_dataOperator_dataOperations_dataPaths():
+def test_dataOperator_DataOperations_data_paths():
 
     with pytest.raises(FileNotFoundError):
-        lst.dataOperations(dataPaths=None)
+        lst.DataOperations(data_paths=None)
 
 
-def test_dataOperator_readProcessedData_fileList():
+def test_dataOperator_ReadProcessedData_file_list():
 
     with pytest.raises(FileNotFoundError):
-        lst.readProcessedData(fileList=None)
+        lst.ReadProcessedData(file_list=None)
 
 
-def test_dataOperator_getRestructuredData_data():
+def test_dataOperator_GetRestructuredData_data():
 
     with pytest.raises(TypeError):
-        lst.getRestructuredData(data=xr.DataArray(np.array([0, 1])))
+        lst.GetRestructuredData(data=xr.DataArray(np.array([0, 1])))
 
 
-def test_dataOperator_getResampled_xrDataArray_none():
+def test_dataOperator_getResampled_xr_data_array_none():
 
     with pytest.raises(TypeError):
-        lst.getResampledData(xrDataArray=np.array([0, 1]))
+        lst.GetResampledData(xr_data_array=np.array([0, 1]))
 
 
-def test_dataOperator_dbsOperations_fileList_none():
+def test_dataOperator_DbsOperations_file_list_none():
 
     with pytest.raises(FileNotFoundError):
-        lst.dbsOperations(fileList=None, varList=["range"])
+        lst.DbsOperations(file_list=None, var_list=["range"])
 
 
-def test_dataOperator_dbsOperations_varList_none():
+def test_dataOperator_DbsOperations_varList_none():
 
     with pytest.raises(KeyError):
-        lst.dbsOperations(fileList=["file_path"], varList=None)
+        lst.DbsOperations(file_list=["file_path"], var_list=None)
