+++ lidarSuit/__init__.py
@@ -23,7 +23,7 @@ except DistributionNotFound:
 from .lstConfig import *
 from .lidar_code import *
 from .windPropRetrieval import *
-from .windPropRetrieval6Beam import *
+from .wind_prop_retrieval_6_beam import *
 from .visualization import *
 from .utilities import *
 from .dataOperator import *
