### 2022-12-16 1ab1af3
Replacing windPropRetrieval by wind_prop_retrieval
