+++ lidarSuit/wind_prop_retrieval.py
@@ -8,8 +8,8 @@ from .filters import Filtering
 from .dataAttributesL1 import LoadAttributes
 from .dataOperator import GetRestructuredData
 
-module_logger = logging.getLogger("lidarSuit.windPropRetrieval")
-module_logger.debug("loading windPropRetrieval")
+module_logger = logging.getLogger("lidarSuit.wind_prop_retrieval")
+module_logger.debug("loading wind_prop_retrieval")
 
 
 class FourierTransfWindMethod:
@@ -38,7 +38,7 @@ class FourierTransfWindMethod:
     def __init__(self, doppler_obs: xr.DataArray):
 
         self.logger = logging.getLogger(
-            "lidarSuit.windPropRetrieval.FourierTransfWindMethod"
+            "lidarSuit.wind_prop_retrieval.FourierTransfWindMethod"
         )
         self.logger.info("creating an instance of FourierTransfWindMethod")
 
@@ -290,7 +290,7 @@ class GetWindProperties5Beam:
     ):
 
         self.logger = logging.getLogger(
-            "lidarSuit.windPropRetrieval.GetWindProperties5Beam"
+            "lidarSuit.wind_prop_retrieval.GetWindProperties5Beam"
         )
         self.logger.info("creating an instance of GetWindProperties5Beam")
 
@@ -536,7 +536,7 @@ class RetriveWindFFT:
     def __init__(self, transfd_data: GetRestructuredData):
 
         self.logger = logging.getLogger(
-            "lidarSuit.windPropRetrieval.FourierTransfWindMethod"
+            "lidarSuit.wind_prop_retrieval.FourierTransfWindMethod"
         )
         self.logger.info("creating an instance of FourierTransfWindMethod")
 
