+++ lidarSuit/__init__.py
@@ -22,7 +22,7 @@ except DistributionNotFound:
 
 from .lstConfig import *
 from .lidar_code import *
-from .windPropRetrieval import *
+from .wind_prop_retrieval import *
 from .wind_prop_retrieval_6_beam import *
 from .visualization import *
 from .utilities import *
