### 2022-12-16 2ad9b2d
Refactor/camel case module name (#77)
* Applying snake case to module name

* Replacing windPropRetrieval6Beam by wind_prop_retrieval_6_beam

* Applying snake case to module name

* Replacing windPropRetrieval by wind_prop_retrieval

* Applying snake case to module name

* Replacing lstConfig by lst_config

* Applying snake case to module name

* Replacing dataOperator by data_operator

* Replacing dataOperator by data_operator

* Applying snake case to module name

* Replacing dataAttributesL1 by data_attributes
