+++ lidarSuit/__init__.py
@@ -20,12 +20,12 @@ except DistributionNotFound:
         )
 
 
-from .lstConfig import *
+from .lst_config import *
 from .lidar_code import *
-from .windPropRetrieval import *
-from .windPropRetrieval6Beam import *
+from .wind_prop_retrieval import *
+from .wind_prop_retrieval_6_beam import *
 from .visualization import *
 from .utilities import *
-from .dataOperator import *
-from .dataAttributesL1 import *
+from .data_operator import *
+from .data_attributes import *
 from .filters import *
