+++ docs/overview.rst
@@ -1,0 +1,98 @@
+==========================
+Scanning strategy overview
+==========================
+
+
+Due to the current transition from fossil fuel-based power plants to renewable sources, wind has become one of the main energy sources. However, to identify the places where wind harvest is suitable for power production, an understanding of how the wind blows in the possible places is needed. Wind Doppler lidars are often used to survey the winds in those places and to estimate turbulence. 
+
+Wind Doppler lidars work by transmitting laser pulses and measuring the shift in frequency of the backscattered signal, the Doppler effect. If the scatterers are moving towards the lidar, the backscattered signal's frequency is higher than that of the transmitted signal and the other way around. 
+
+There are different scanning strategies for estimating wind profiles using Wind Doppler lidars, but two are used more often: the 6-beam and the DBS.
+
+-------
+6-beam:
+-------
+
+Reynolds stress tensor calculation
+----------------------------------
+
+The 6-beam is a scanning strategy that is idealised for retrieving turbulence. This method measures the radial velocity at five azimuths (0, 72, 144, 216, 288º) with a certain elevation angle and one additional observation at 90º elevation. The cartoon below illustrates the geometry of this strategy.
+
+Cartoon:
+
+`Sathe et al.,2015 <https://amt.copernicus.org/articles/8/729/2015/>`_  suggested that the variance of each radial observation can be expressed in terms of variances and co-variances of u (zonal wind), v (meridional wind), and w (vertical wind), as indicated below. :math:`\theta` and :math:`\phi` are the azimuth and elevation angles, respectively.
+
+.. math::
+
+    var(v_{\theta}) = var(u)cos^2(\phi)sin^2(\theta) + var(v)cos^2(\phi)cos^2(\theta) + var(w)sin^2(\phi) \\
+    + 2cov(u,v)cos^2(\phi)cos(\theta)sin(\theta) + 2cov(u,w)cos(\phi)sin(\phi)sin(\theta) \\ 
+    + 2 cov(v,w)cos(\phi)sin(\phi)cos(\theta)
+
+
+Combining all the equations for each one of the six observations makes it possible to solve this equation system for the variances and co-variances of u, v, and W, which are the Reynolds stress tensor components.  
+
+.. math::
+
+    \stackrel{\Sigma}{
+    \begin{bmatrix} var(u)\\ var(v)\\ var(w)\\ 
+                    cov(u,v)\\ cov(u,w)\\ cov(v,w)\end{bmatrix}} = 
+    \stackrel{S}{M^{-1}\begin{bmatrix} var(v_{0})\\ var(v_{72})\\ var(v_{144})\\
+                    var(v_{216})\\ var(v_{288})\\ var(v_{zenith})
+    \end{bmatrix}}
+
+
+
+Wind speed and direction estimation
+-----------------------------------
+
+Even though the 6-beam strategy makes it possible to calculate the Reynolds stress tensor directly, the horizontal wind speed and direction estimation is not straightforward. One of the possible ways to estimate the wind speed and direction is using the Fast Fourier Wind Vector Algorithm (FFWVA) developed by `Ishwardat, 2017 <http://resolver.tudelft.nl/uuid:a659654b-e76a-4513-a656-ecad761bdbc8>`_. 
+
+The FFWVA uses Fast Fourier Transform (FFT) digital signal-processing algorithms to decompose the radial Doppler observations in terms of amplitude and phase of their harmonic frequencies. The wind speed :math:`V_{s}` and direction :math:`V_{d}` can be calculated using the amplitude and phase from the first harmonic, as indicated below.
+
+.. math::
+    a + bi = FFT(v_{\theta})|_{1st}
+    
[…]
