### 2022-12-20 fa57376
Adding API to the sidebar menu
