### 2022-12-20 fcbeb25
Adding a description of the diffent scanning aproachs
