+++ docs/readme.rst
@@ -1,1 +1,2 @@
+.. _intro:
 .. include:: ../README.rst
