+++ docs/readme.rst
@@ -1,2 +1,5 @@
 .. _intro:
 .. include:: ../README.rst
+
+.. note::
+    All publicly available classes are listed on the :any:`API reference <api>`.
