+++ tox.ini
@@ -10,7 +10,7 @@ python =
 [testenv:flake8]
 basepython = python
 deps = flake8
-commands = flake8 lidar_suit tests
+commands = flake8 lidarSuit tests
 
 [testenv]
 setenv =
