### 2022-12-21 8537a2b
Adding contribution section to the readthedocs readme
