+++ docs/readme.rst
@@ -3,3 +3,9 @@
 
 .. note::
     All publicly available classes are listed on the :any:`API reference <api>`.
+
+------------
+Contribution
+------------
+
+Any contribution to further development is appreciated, and guidelines for the contributions are available in the :any:`contributing section <contributing>`. To support further development, sample datasets from different scanning strategies are available on Zenodo, and they can be found at https://doi.org/10.5281/zenodo.7312960.
