+++ CONTRIBUTING.rst
@@ -4,6 +4,7 @@
 Contributing
 ============
 
+
 Contributions are welcome, and they are greatly appreciated! Every little bit
 helps, and credit will always be given.
 
