+++ README.rst
@@ -35,6 +35,15 @@ Documentation
 The lidarSuit's documentation is available at https://lidarsuit.readthedocs.io, and an introductory set of notebooks are available at https://nbviewer.org/github/jdiasn/lidarSuit/tree/main/docs/notebooks/.
 
 
+
+.. warning::
+
+    Beware that between versions 0.1.6 and 0.2.0, the package underwent significant refactoring. Now the classes' names
+    follow the Pascal case, while module names, functions and attributes follow the snake case. Codes developed using the previous
+    version will need revision.
+
+
+
 ----------
 Suggestion
 ----------
