### 2022-12-23 3457c96
Upgrade/documentation (#78)
* Adding overview and updating api

* Adding new the overview to the index

* Reformating overview.rst

* Correcting typos #53

* Removing multiples links to the introduction #46

* Adding API to the sidebar menu

* Adding a description of the diffent scanning aproachs

* Extending the usaged description #46 #53

* Adding a tag to the introduction

* Adding figures

* Adjusting the introduction

* Package name consistence

* Adding contribution section to the readthedocs readme

* Moving notebooks to the root

* Creating notebooks galleries

* Updating links

* Corr
[…]
