+++ docs/usage.rst
@@ -39,3 +39,112 @@ Whenever the class lst.RetriveWindFFT() is used to retrieve wind, lidarSuit will
     'You do not have a config file yet'
     'a temporary config file was generated'
     'See the documentation for generating it'
+
+
+
+-------------------
+Data pre-processing
+-------------------
+
+
+The NetCDF files produced by the WindCube operating system while following the 6-beam or the DBS can differ from each other. The methods for retrieving wind also differ from each other. For those reasons, the lidarSuit has different pre-processing classes to handle those files and create the requirements for deriving wind.
+
+
+
+6-beam data pre-processing
+--------------------------
+
+
+The 6-beam data is pre-processed by the DataOperations class. It merges all data, grouping the observations into zenith pointing and slanted. To use this class, the user only needs to pass a list of file paths. The merged data set can be accessed using the merged_data attribute. After pre-processing the data, one can also save the pre-processed data as NetCDF and continue with the wind retrieval or turbulence estimation afterwards.
+
+.. code-block:: python
+
+    >>> merged_ds = lidarSuit.DataOperations(file_list).merged_data
+    >>> merged_ds.to_netcdf(output_file_path)
+
+
+DBS data pre-processing
+--------------------------
+
+The pre-processing of the DBS data is handled by the DbsOperations class. This class needs a list of file paths and a list of variables that will be merged. While executing, it adds a mean time variable to identify all observations from a complete DBS scan. After using this class, the user can access the merged data by calling the merged_ds attribute, which can also be saved as a NetCDF file to continue with the wind retrieval afterwards.
+
+
+.. code-block:: python
+
+    >>> var_list = ['azimuth', 'elevation', 'radial_wind_speed',
+    >>>             'radial_wind_speed_status', 'measurement_height', 'cnr']
+    >>> merged_ds = lst.DbsOperations(file_list, var_list).merged_ds
+    >>> merged_ds.to_netcdf(output_file_path)
+
+
+---------------
+Retrieving wind
+---------------
+
+lidarSuit is able to retrieve wind from the 6-beam data and also from the DBS data.
+
+6-beam data wind retrieval
+--------------------------
+
+To retrieve the wind speed and direction from the 6-beam data, the user needs to prepare the dataset for it. It is made by using the GetRestructuredData class. This class groups the slanted data by their azimuth and creates an azimuth dimension to store it. Once it is done, the output object from GetRestructuredData class contains all the information needed for retrieving wind profiles using the RetriveWindFFT class.
+
+The RetriveWindFFT class applies the fast Fourier transform (FFT) along the azimuthal dimension of the dataset for each complete scanning cycle and derives the wind speed and direction from that. A short description of the FFT-based wind retrieval can be found in the :any:`6-beam <six-beam>` section. Below you can see how to retrieve winds from the 6-data. The wind_obj has an attribute wind_prop (a xarray dataset) where the wind profiles are stored.
+
+
+.. code-block:: python
+
+    >>> restruct_data = lst.GetRestructuredData(merged_ds)
+    >>> wind_obj = lst.RetriveWindFFT(restruct_data)
+
+
+A notebook example combining all steps for retrieving wind can be found in the list of `notebooks examples <examples/merging_6beam_rendered.html>`_.  You can run the same example online by clicking on the binder badge listed in the package :any:`introduction <intro>`.
+
+
+DBS data wind retrieval
+-----------------------
+
[…]
