+++ docs/notebooks.rst
@@ -1,0 +1,10 @@
+==================
+Rendered notebooks
+==================
+
+.. nbgallery::
+    examples/quicklooks_rendered.ipynb
+    examples/merging_6beam_rendered.ipynb
+    examples/turbulence_6beam_data_rendered.ipynb
+    examples/dbs_scans_rendered.ipynb
+    examples/reading_long_dbs_rendered.ipynb
