### 2022-12-23 4abf597
Small documentation issue (#80)
* Adding the Rendered notebooks section

* Updating binder
