### 2022-12-24 45907cb
Removing old commented code and flake8
