### 2022-12-24 fbf473f
Removing old commented code and flake8
