### 2022-12-24 911c6c9
Removing old commented code and flake8
