### 2022-12-24 e910562
formated string
