### 2022-12-24 3ba6c4f
Removing old commented code and flake8
