### 2022-12-24 ef6e69e
Removing old commented code and flake8
