### 2022-12-26 457024a
Update/removing old code (#82)
* Updating style (flake8)

* Removing old commented code and flake8

* Removing old commented code and flake8

* Removing old commented code and flake8

* Removing extra space

* Checking with flake8

* Removing old commented code and flake8

* formated string

* Removing old commented code and flake8

* Removing old commented code and flake8
