### 2022-12-27 29a6ceb
Adding an output example
