### 2022-12-27 4d2b073
Removing not needed sentence
