### 2022-12-27 2d89742
Reviewing the manuscript  (#83)
* Adding an output example

* Removing not needed sentence
