+++ joss/paper.bib
@@ -82,3 +82,28 @@
   number={4},
   pages={76-82},
   doi={10.1109/38.56302}}
+
+
+
+@article{diasneto2022b,
+	author = {Dias Neto, J. and Nuijens, L. and Unal, C. and Knoop, S.},
+	doi = {10.5194/essd-2022-268},
+	journal = {Earth System Science Data Discussions},
+	pages = {1-30},
+	title = {Combined Wind Lidar and Cloud Radar for Wind Profiling},
+	url = {https://essd.copernicus.org/preprints/essd-2022-268/},
+	volume = {2022},
+	year = {2022},
+	bdsk-url-1 = {https://essd.copernicus.org/preprints/essd-2022-268/},
+	bdsk-url-2 = {https://doi.org/10.5194/essd-2022-268}}
+
+
+@BOOK{Stull2003,
+  title     = "An Introduction to Boundary Layer Meteorology ",
+  author    = "Stull, Roland B.",
+  publisher = "Springer Dordrecht",
+  edition   =  "1",
+  year      =  "2003",
+  address   = "Dordrecht",
+  isbn = "978-90-277-2769-5",
+}
