### 2022-12-28 146bfac
Adding a diagram of the package structure (#84)
