### 2022-12-28 5db2308
Adding test for the sixbeammethod imput
