### 2022-12-29 ee32761
Updating module name in tests
