### 2022-12-29 08ae6e0
Let's identify the test reference by a generic key
Later we might be interested in specific file or group of files, so
let's call it key.
