### 2022-12-29 b1bcee7
Prototype of pre-processing (#60)
* Prototype of pre-processing

@jdiasn, could you point me to an example of elevation=90 to guide this
development, please? It is not clear to me if we can have a 90 degrees
case in the same file with a non 90 degres elevation.

* fix: round is int

* Pre-processing sweep

@jdiasn, based on what we talked about, does it make sense?

* Naming new function as suggested by @jdiasn

* Initializing a support module to handle testing datasets

* Moving support functions to its own module

* New support function lidarsuitrc()

* New support function for tests: sample
[…]
