### 2022-12-29 120bf38
Refactor/filtering (#59)
* A simple function to filter based on the status

* Prottype of filter_snr()

* Formated string for simplicity
