### 2022-12-29 6f14ba7
Six beam method tests
