### 2022-12-29 e71d803
Adding test for the fft wind retrieval method
