### 2022-12-29 55190c4
Activating phase and direction tests
