### 2022-12-29 918b23b
Activating tests not related to the velocity field
The new version of xrft is leading to a wrong estimation
of wind speed
