### 2022-12-29 f8fe053
Activating the wind speed related tests
