### 2022-12-29 5caee97
Adding test for the fft wind retrieval method (#86)
* Adding test for the fft wind retrieval method

* Deactivating some tests

* Activating phase and direction tests

* Activating tests not related to the velocity field

The new version of xrft is leading to a wrong estimation
of wind speed

* Fixing xrft compatibility problem

* Activating the wind speed related tests
