### 2022-12-29 1c2df7e
Testing if variables exist
