### 2022-12-29 b4828d8
Testing the variables values and global attributes
