### 2022-12-29 1ad3d42
Feat/test retrieve wind (#87)
* Testing if variables exist

* Testing the variables values and global attributes
