### 2022-12-30 4dfadc4
Testing the variables dimension
