### 2022-12-30 9e01ece
Adding tests for the variables values
