### 2022-12-30 a07bdf2
Testing comp u value
