### 2022-12-30 c02625e
Alternative way of testing comp u value
