### 2022-12-30 6ac0f8c
Testing wind speed value
