### 2022-12-30 5f46406
Rounding all velocity values before testing
