### 2022-12-30 5cecd69
Adding test for wind direction and vertical wind values
