### 2022-12-30 405d8ee
Testing GetWindProperties5Beam class (#88)
* Testing the variables dimension

* Adding tests for the variables values

* Testing comp u value

* Alternative way of testing comp u value

* Setting value to float

* Testing comp v

* Testing wind speed value

* Rounding the value for testing

* Rounding all velocity values before testing

* Adding test for wind direction and vertical wind values
