### 2022-12-30 382cada
Testing GetRestructuredData class (#89)
* Testing size and values from range_non_90

* Testing dimensions from data_transf

* Testing elevation and azimuth values from data_transf

* Testing dimensions from data_transf_90

* Testing data_transf_90 name and dimension size

* Testing broken input dataset
