### 2022-12-30 jdiasn/lidarSuit#48 @jdiasn
@tgoelles, thanks for raising this issue. It was indeed an inadequate naming standard. It was addressed and all related PRs were linked here. After those modifications, the current score is around 8.74. I'm closing this issue but please feel free to open it again if I missed anything. Thanks.
