### 2022-12-30 afbbbca
Replacing lidarSuit with lidarwind: lst_config
