+++ lidarwind/lst_config.py
@@ -13,8 +13,8 @@ class Configurations:
 
     Parameters
     ----------
-    lst : object
-        an instance of the lidarSuit package
+    lidarwind : object
+        an instance of the lidarwind package
 
     """
 
@@ -31,7 +31,7 @@ class Configurations:
 
     def load_version(self, lst):
         """
-        It identifies the lidarSuit version
+        It identifies the lidarwind version
         and writes it to the configuration file
 
         Parameters
@@ -122,15 +122,15 @@ class Configurations:
 
         return self
 
-    def load_reference(self, reference="Generated by lidarSuit version: {0}"):
+    def load_reference(self, reference="Generated by lidarwind version: {0}"):
         """
-        It loads the lidarSuit's version used for
+        It loads the lidarwind's version used for
         processing the data
 
         Parameters
         ----------
         reference : str
-            lidarSuit version used to process the data
+            lidarwind version used to process the data
 
         """
 
