+++ lidarwind/__init__.py
@@ -20,7 +20,7 @@ except DistributionNotFound:
         )
 
 
-from .lst_config import *
+from .lidarwind_config import *
 from .lidar_code import *
 from .wind_prop_retrieval import *
 from .wind_prop_retrieval_6_beam import *
