### 2022-12-30 c3ed8bd
Replacing lst_config with lidarwind_config: __init__
