+++ lidarwind/data_attributes.py
@@ -15,7 +15,7 @@ class LoadAttributes:
     ----------
     data : xarray.Dataset
         A xarray dataset containing the Level 1 data
-        generated by the lidarsuit
+        generated by the lidarwind
 
     Returns
     -------
