+++ lidarwind/wind_prop_retrieval_6_beam.py
@@ -26,7 +26,7 @@ class SixBeamMethod:
     ----------
     data : object
         an instance of the object generated by the
-        lst.GetRestructuredData()
+        lidarwind.GetRestructuredData()
 
     freq : int
         number of profiles used to calculate
