### 2022-12-30 e0c4571
Removing old commented code
