### 2022-12-30 7a95846
Replacing lidarSuit with lidarwind: test_get_restructured_data
