### 2022-12-30 0d16d03
Replacing lidarSuit with lidarwind: test_wind_prop_retrieval_6_beam
