+++ tests/test_wind_prop_retrieval_6_beam.py
@@ -2,7 +2,7 @@ import pytest
 import numpy as np
 import xarray as xr
 
-import lidarSuit as lst
+import lidarwind as lst
 
 
 def test_six_beam_method_input():
