+++ tests/test_IO.py
@@ -1,6 +1,6 @@
 import pytest
 
-import lidarSuit as lst
+import lidarwind as lst
 from .data import data_filenames, get_sample_data
 
 
