### 2022-12-30 facf6a2
Replacing lidarSuit with lidarwind: test_dataOperator
