+++ tests/test_dataOperator.py
@@ -2,8 +2,8 @@ import pytest
 import numpy as np
 import xarray as xr
 
-from lidarSuit.data_operator import wc_fixed_preprocessing
-import lidarSuit as lst
+from lidarwind.data_operator import wc_fixed_preprocessing
+import lidarwind as lst
 
 from .data import sample_dataset
 
