### 2022-12-30 f0e8b82
Partialy replacing lidarSuit with lidarwind: pyproject.toml
