+++ CONTRIBUTING.rst
@@ -16,7 +16,7 @@ Types of Contributions
 Report Bugs
 ~~~~~~~~~~~
 
-Report bugs at https://github.com/jdiasn/lidarSuit/issues.
+Report bugs at https://github.com/jdiasn/lidarwind/issues.
 
 If you are reporting a bug, please include:
 
@@ -46,7 +46,7 @@ articles, and such.
 Submit Feedback
 ~~~~~~~~~~~~~~~
 
-The best way to send feedback is to file an issue at https://github.com/jdiasn/lidarSuit/issues.
+The best way to send feedback is to file an issue at https://github.com/jdiasn/lidarwind/issues.
 
 If you are proposing a feature:
 
@@ -58,17 +58,17 @@ If you are proposing a feature:
 Get Started!
 ------------
 
-Ready to contribute? Here's how to set up `lidarSuit` for local development.
+Ready to contribute? Here's how to set up `lidarwind` for local development.
 
-1. Fork the `lidarSuit` repo on GitHub.
+1. Fork the `lidarwind` repo on GitHub.
 2. Clone your fork locally::
 
-    $ git clone git@github.com:your_name_here/lidarSuit.git
+    $ git clone git@github.com:your_name_here/lidarwind.git
 
 3. Install your local copy into a virtualenv. Assuming you have virtualenvwrapper installed, this is how you set up your fork for local development::
 
-    $ mkvirtualenv lidarSuit
-    $ cd lidarSuit/
+    $ mkvirtualenv lidarwind
+    $ cd lidarwind/
     $ python setup.py develop
 
 4. Create a branch for local development::
@@ -80,7 +80,7 @@ Ready to contribute? Here's how to set up `lidarSuit` for local development.
 5. When you're done making changes, check that your changes pass flake8 and the
    tests, including testing other Python versions with tox::
 
-    $ flake8 lidarSuit tests
+    $ flake8 lidarwind tests
     $ python setup.py test or pytest
     $ tox
 
@@ -104,7 +104,7 @@ Before you submit a pull request, check that it meets these guidelines:
    your new functionality into a function with a docstring, and add the
    feature to the list in README.rst.
 3. The pull request should work for Python 3.5, 3.6, 3.7 and 3.8, and for PyPy. Check
-   https://travis-ci.com/jdiasn/lidarSuit/pull_requests
+   https://travis-ci.com/jdiasn/lidarwind/pull_requests
    and make sure that the tests pass for all supported Python versions.
 
 Tips
@@ -112,7 +112,7 @@ Tips
 
 To run a subset of tests::
 
-$ pytest tests.test_lidarSuit
+$ pytest tests.test_lidarwind
 
 
 Deploying
