### 2022-12-30 3d5aeba
Partially replacing lidarSuit with lidarwind:README.rst
