+++ tox.ini
@@ -10,7 +10,7 @@ python =
 [testenv:flake8]
 basepython = python
 deps = flake8
-commands = flake8 lidarSuit tests
+commands = flake8 lidarwind tests
 
 [testenv]
 setenv =
