### 2023-01-03 #4852 @elbeejay
@jdiasn, thanks for working to revise your package in accordance with reviewer suggestions. I've taken a look at some of the outstanding issues and wanted to just add my support for some of the suggestions from our reviewers.

Firstly, it looks like you have done a lot internally to change variable naming from camel-case to snake-case. I get the impression you are also considering changing the package name (https://github.com/jdiasn/lidarSuit/issues/49)? I think this is something you should consider, as it was also mentioned in https://github.com/jdiasn/lidarSuit/issues/53 - this is a change I'd like to see before we ask our reviewers to look at the package again.

I agree that the storage of sample data on Google Drive is not reliable and can be avoided. Especially as free archival services (e.g., Zenodo) are available (https://github.com/jdiasn/lidarSuit/issues/53). I think the `pooch` package is a great suggestion, although not required. I think you could upload files to an archive and download them similarly to what you set up for Google Drive and that would be just fine, but I do agree that moving off of Google Drive for storage of sample data would be the recommended course of action.
