### 2023-01-03 #4852 @jdiasn
[elbeejay](https://github.com/elbeejay), I think I completed 90% of the modifications. I have already started the name modification of the package, and I have a PR dedicated to it (jdiasn/lidarSuit#90).  To be consistent, I will also rename the git hub repository.  Do you think that changing the repository name will cause problems with the submission?

I also agree about the storage. I moved the sample data to Zenodo already, and pooch is already used in the notebooks examples for accessing the sample data.

After finishing the modification, should I just close the issues, or should I ask the reviewers for it?
