### 2023-01-04 #4852 @elbeejay
> After finishing the modifications, should I just close the issues, or should I ask the reviewers for it?

Either way is fine. Personally I'd ask reviewers so that they can close the issue and resolve the outstanding checkboxes in their checklist here at the same time.
