### 2023-01-08 jdiasn/lidarwind#49 @jdiasn
@tgoelles, the package was renamed, and the new name is lidarwind. I reviewed the entire package, and all wrong usage of the camel case was fixed.  I think I completed all issues from this PR. If you agree with it you can close this PR.
