### 2023-01-08 jdiasn/lidarwind#46 @jdiasn
@bgailleton, I believe I addressed all issues listed here. The notebooks were expanded, and a section of rendered notebooks was added to the documentation (https://lidarwind.readthedocs.io/en/latest/notebooks.html). The usage section was expanded to indicate how to use the different classes for pre-processing the data and retrieving wind. The 3 times reference to the intro was solved. In the introduction of the documentation, the user is notified of the availability of an API and a section of rendered docstrings was added to the documentation as well (https://lidarwind.readthedocs.io/en/latest/modules.html#). The paper was updated and expanded following the suggestion from another review.

If you think all points listed in this issue were addressed, you can close this issue.
