### 2023-01-08 jdiasn/lidarwind#53 @jdiasn
@acocac,  I believe I addressed all issues listed here. The notebooks were expanded to improve the usage description, and a section of rendered notebooks was added to the documentation (https://lidarwind.readthedocs.io/en/latest/notebooks.html). The usage section of the documentation was expanded to explain how the package can be used for pre-processing the data and retrieving winds from the different scanning strategies. A section introducing the different scanning strategies and how the wind is retrieved was also added to the documentation (https://lidarwind.readthedocs.io/en/latest/overview.html). The package was renamed. The sample data was moved to Zenodo, and the notebook examples are using pooch to download the data.  The wind field retrieved from a particular day was added to the paper, but the diagram was only added to the documentation.

If you think all points listed in this issue were addressed, you can close this issue.
