### 2023-01-08 jdiasn/lidarwind#49 @tgoelles
Yes the new name is good and camel case is gone.

https://pypi.org/classifiers/

Just an additional hint to improve the pypi page:

https://pypi.org/classifiers/

and use the readme in the long_description in setup.py
