### 2023-01-08 jdiasn/lidarwind#49 @jdiasn
Thank you, @tgoelles! I will look at those additional points to further improve the package.
