### 2023-01-10 jdiasn/lidarwind#54 @jdiasn
Hi @tgoelles, I believe I addressed all points you listed here.
I managed to add the docstrings to the readthedocs (https://lidarwind.readthedocs.io/en/latest/modules.html). That was a really nice suggestion.
