### 2023-01-12 jdiasn/lidarwind#54 @tgoelles
Hi,

I would still add nbval to the tests
https://pypi.org/project/nbval/

But if you dont want to its also fine.
