### 2023-01-12 #4852 @elbeejay
Great, @bgailleton indicated support for publication above, and it looks like @tgoelles has kept track of the revisions and completed their checklist accordingly. **To be explicit, @tgoelles do you consider this submission now ready for publication?**

@acocac now that @jdiasn has completed revisions in response to reviewer comments I wanted to ping you and ask that you revisit this submission and go over your review comments at your earliest convenience.

Thanks all!
Jay
