### 2023-01-14 jdiasn/lidarwind#53 @acocac
@jdiasn thanks for addressing my comments. All implemented changes look good to me. It's ok to include the diagram in the documentation instead of the manuscript. Also, the figure of wind field retrieval is a great example and it was nicely explained with a plain language.

My only comment is regarding the Binder. Considering the relevance of this resource, I suggest to update the following text within [DBS data wind retrieval](https://lidarwind.readthedocs.io/en/latest/usage.html?highlight=binder#dbs-data-wind-retrieval) and [6-beam data wind retrieval](https://lidarwind.readthedocs.io/en/latest/usage.html?highlight=binder#beam-data-wind-retrieval) notebooks. Also the Binder doens't work due to conflicts with the lidarversion:

```
ERROR: Could not find a version that satisfies the requirement lidarwind==0.1.6 (from versions: 0.2.1)
ERROR: No matching distribution found for lidarwind==0.1.6
```

I strongly encourage you to address the issue. Binder would facilitate potential users to run and explore your package. If it's ok I'll open a new issue and then you close it at your earliest convenience.
