### 2023-01-14 jdiasn/lidarwind#96 @acocac
Binder fails to launch
* LIDAR Suit version: 0.1.6
* Python version: Public binder
* Operating System: Public binder

### Description

When run Binder in _nbviewer_ it returns an error related to the version of lidarwind.

### What I Did

```
ERROR: Could not find a version that satisfies the requirement lidarwind==0.1.6 (from versions: 0.2.1)
ERROR: No matching distribution found for lidarwind==0.1.6
```
