### 2023-01-14 #4852 @acocac
@elbeejay I went through my suggested changes in https://github.com/jdiasn/lidarwind/issues/53. @jdiasn has implemented most of them so I completed my checklist. I only left a suggestion to fix the launch of Binder by opening a separate issue, however this doesn't affect my fully support to publish the paper (:
