### 2023-01-15 jdiasn/lidarwind#96 @jdiasn
@acocac, I fixed the binder issue. Now it is working again.
