### 2023-01-17 jdiasn/lidarwind#96 @acocac
> @acocac, I fixed the binder issue. Now it is working again.

@jdiasn thanks for fixing the lidarwind version, it seems the Binder still have some issues even I rerun on an private tab in my browser. Can you double-check?

```
# >>>>>>>>>>>>>>>>>>>>>> ERROR REPORT <<<<<<<<<<<<<<<<<<<<<<

    Traceback (most recent call last):
      File "/srv/conda/lib/python3.10/site-packages/conda/exceptions.py", line 1114, in __call__
        return func(*args, **kwargs)
      File "/srv/conda/lib/python3.10/site-packages/conda_env/cli/main.py", line 80, in do_call
        exit_code = getattr(module, func_name)(args, parser)
      File "/srv/conda/lib/python3.10/site-packages/conda_env/cli/main_update.py", line 127, in execute
        result[installer_type] = installer.install(prefix, specs, args, env)
      File "/srv/conda/lib/python3.10/site-packages/mamba/mamba_env.py", line 55, in mamba_install
        index = load_channels(pool, channel_urls, repos, prepend=False)
      File "/srv/conda/lib/python3.10/site-packages/mamba/utils.py", line 129, in load_channels
        index = get_index(
      File "/srv/conda/lib/python3.10/site-packages/mamba/utils.py", line 95, in get_index
        api.create_cache_dir(str(pkgs_dirs.first_writable_path))
    RuntimeError: filesystem error: cannot create directories: No space left on device [cache]

`$ /srv/conda/bin/mamba update -p /srv/conda/envs/notebook --file environment.yml`

  environment variables:
                 CIO_TEST=<not set>
  CONDA_AUTO_UPDATE_CONDA=false
        CONDA_DEFAULT_ENV=/srv/conda/envs/notebook
                CONDA_DIR=/srv/conda
           CONDA_PLATFORM=linux-64
               CONDA_ROOT=/srv/conda
           CURL_CA_BUNDLE=<not set>
                     PATH=/home/jovyan/.local/bin:/home/jovyan/.local/bin:/srv/conda/envs/notebo
                          ok/bin:/srv/conda/bin:/srv/npm/bin:/usr/local/sbin:/usr/local/bin:/usr
                          /sbin:/usr/bin:/sbin:/bin
       REQUESTS_CA_BUNDLE=<not set>
            SSL_CERT_FILE=<not set>

     active environment : None
       user config file : /home/jovyan/.condarc
 populated config files : /srv/conda/.condarc
          conda version : 4.13.0
    conda-build version : not installed
         python version : 3.10.8.final.0
       virtual packages : __linux=5.15.0=0
                          __glibc=2.27=0
                          __unix=0=0
                          __archspec=1=x86_64
       base environment : /srv/conda  (writable)
      conda av data dir : /srv/conda/etc/conda
  conda av metadata url : None
           channel URLs : https://conda.anaconda.org/conda-forge/linux-64
                          https://conda.anaconda.org/conda-forge/noarch
                          https://repo.anaconda.com/pkgs/main/linux-64
                          https://repo.anaconda.com/pkgs/main/noarch
                          https://repo.anaconda.com/pkgs/r/linux-64
                          https://repo.anaconda.com/pkgs/r/noarch
          package cache : /srv/conda/pkgs
                          /home/jovyan/.conda/pkgs
       envs directories : /srv/conda/envs
                          /home/jovyan/.conda/envs
               platform : linux-64
             user-agent : conda/4.13.0 requests/2.28.1 CPython/3.10.8 Linux/5.15.0-56-generic ubuntu/18.04.6 glibc/2.27
                UID:GID : 1000:1000
             netrc file : None
           offline mode : False

An unexpected error has occurred. Conda has prepared the above report.

time: 0.798
```
