### 2023-01-18 jdiasn/lidarwind#96 @jdiasn
@acocac I do not get why you have this error. Every time I lunch the binder, it works for me.  Have a look at this screenshot.
