### 2023-01-18 jdiasn/lidarwind#96 @acocac
You're right, it works now! Feel free to close the issue.
