### 2023-01-18 #4852 @jdiasn
The issue jdiasn/lidarwind#96 is solved.
