### 2023-01-18 #4852 @elbeejay
@jdiasn great - thanks for making all of these revisions. Big thank you to our reviewers as well, @tgoelles, @acocac, @bgailleton your constructive comments resulted in a number of changes to this package that I think have improved it.

At this time @jdiasn, can you do the following:
- [ ] (if necessary) Re-release the package on PyPi so that there is a pip-installable version of the package corresponding to what is going to be the JOSS-publication version
