### 2023-01-31 #4852 @jdiasn
@elbeejay, I reviewed the list of points you mentioned, and I believe I completed them.

Package tag: https://github.com/jdiasn/lidarwind/releases/tag/v0.2.2
Zenodo archive: https://zenodo.org/record/7591728#.Y9mMkhPMI-Q
DOI: 10.5281/zenodo.7591728
PyPI version: https://pypi.org/project/lidarwind/0.2.2/
