### 2023-02-06 #4852 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
